### 2023-07-10 #5642 @editorialbot
[PRE REVIEW]: simChef: High-quality data science simulations in R
**Submitting author:** @jpdunc23 (<a href="http://orcid.org/0000-0003-3297-681X">James Duncan</a>)
**Repository:** https://github.com/Yu-Group/simChef
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.3
