+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   parallel: parallel.html
   simChef: simChef.html
   simulation-directory-setup: simulation-directory-setup.html
-last_built: 2023-07-03T18:26Z
+last_built: 2023-08-14T18:09Z
 urls:
   reference: https://yu-group.github.io/simChef/reference
   article: https://yu-group.github.io/simChef/articles
