### 2023-08-15 #5642 @dhhagan
No worries @brunomontezano - thanks!

Sorry for the dropping the ball a bit here, @jpdunc23 - I've pinged some potential reviewers offline, but seems everyone is away from August.

👋 @rcannood @simonpcouch @stemangiola Are you able and interested in reviewing this manuscript and code for JOSS?
