### 2023-08-15 #5642 @rcannood
Looks like a fun project / paper to review, so sure!

@dhhagan In light of recent developments, does JOSS have any guidelines or policies w.r.t. LLMs? Are there any expectations of authors of reviewers while submitting / reviewing a manuscript?
