### 2023-08-16 7014422
partial update of simChef vignette
