+++ vignettes/simChef_2methods.Rmd
@@ -45,7 +45,7 @@ grammar](https://design.tidyverse.org/). The core features of `simChef` include:
 
 To highlight the ease of running simulation studies using `simChef`, we begin with a basic example usage and defer details to Section \@ref(writing-your-own-simulation-experiment).
 
-> **Simulation Example:** Suppose, for concreteness, that we would like to evaluate the prediction accuracy of two methods (linear regression and random forests) under a linear data-generating process across different signal-to-noise ratios.
+> **Simulation Example:** Suppose, for concreteness, that we would like to evaluate the prediction accuracy of two methods (linear regression and random forests) under two different underlying data-generating processes (a linear and non-linear response function) and across different signal-to-noise ratios.
 
 At its core, `simChef` breaks down a simulation experiment into four components:
 
@@ -58,7 +58,7 @@ Using these components, users can easily run a simulation experiment in six simp
 
 ### Step 1. Specify DGP, method, evaluation, and visualization functions of interest. {-}
 
-Recall that in the given simulation example, we would like to study two methods (linear regression and random forests) under a linear DGP. We code this DGP (`linear_dgp_fun`) and methods (`linear_reg_fun` and `rf_fun`) via custom functions below. 
+Recall that in the given simulation example, we would like to study two methods (linear regression and random forests) under two different DGPs (a linear and non-linear response function). We code these DGPs (`linear_dgp_fun` and `nonlinear_dgp_fun`) and methods (`linear_reg_fun` and `rf_fun`) via custom functions below. 
 
 ```{r dgp-funs}
 #' Linear Data-Generating Process
@@ -78,8 +78,31 @@ linear_dgp_fun <- function(n_train, n_test, p, beta, noise_sd) {
   data_list <- list(
     X_train = X[1:n_train, , drop = FALSE],
     y_train = y[1:n_train],
-    X_test = X[(n_train + 1):n, , drop = FALSE],
-    y_test = y[(n_train + 1):n]
+    X_test = X[(n_train + 1):n_test, , drop = FALSE],
+    y_test = y[(n_train + 1):n_test]
+  )
+  return(data_list)
+}
+
+#' Non-linear Data-Generating Process
+#' 
+#' @description Generate training and test data according to the following 
+#'   non-linear data-generating process: y = 1(X > 0)*beta + noise.
+#' 
+#' @param n_train Number of training samples.
+#' @param n_test Number of training samples.
+#' @param p Number of features.
+#' @param beta Coefficient vector in linear regression function.
+#' @param noise_sd Standard deviation of additive noise term.
+nonlinear_dgp_fun <- function(n_train, n_test, p, beta, noise_sd) {
+  n <- n_train + n_test
+  X <- matrix(rnorm(n * p), nrow = n, ncol = p)
+  y <- (X > 0) %*% beta + rnorm(n, sd = noise_sd)
+  data_list <- list(
+    X_train = X[1:n_train, , drop = FALSE],
+    y_train = y[1:n_train],
+    X_test = X[(n_train + 1):n_test, , drop = FALSE],
+    y_test = y[(n_train + 1):n_test]
   )
   return(data_list)
 }
@@ -129,6 +152,11 @@ linear_dgp <- create_dgp(
   # additional named parameters to pass to .dgp_fun()
   n_train = 200, n_test = 200, p = 2, beta = c(1, 0), noise_sd = 1
 )
+nonlinear_dgp <- create_dgp(
[…]
