+++ vignettes/simChef_full.Rmd
@@ -1,0 +1,910 @@
+---
+title: "Getting started with simChef - Full Guide"
+output:
+  rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{simChef Vignette}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  # disable R chunk evaluation when using macOS in GH Actions
+  # see https://github.com/tidyverse/ggplot2/issues/2252#issuecomment-1006713187
+  # and https://github.com/lcolladotor/biocthis/issues/27 and
+  # https://github.com/bodkan/slendr/commit/f0f5ae18452cc9df5d151874e41b0b8fd5f29aa2#
+  eval = Sys.getenv("RUNNER_OS") != "macOS"
+)
+
+library(simChef)
+
+set.seed(12345)
+
+# remove old cached results to start fresh
+for (fname in list.files(file.path("results", "Linear Regression Experiment"),
+                         pattern = ".rds", recursive = T, full.names = T)) {
+  file.remove(fname)
+}
+```
+
+# Overview
+
+The goal of `simChef` is to seamlessly and efficiently run simulation experiments using a simple grammar. The results of these simulation experiments can also be conveniently viewed in an organized and interactive browser (or html file) (e.g., [here](../linear_regression_output.html)). The basic usage of `simChef` can be summarized as follows:
+
+1. Create individual parts of the simulation experiment recipe
+    a. Create DGP(s) (data-generating processes) via `create_dgp()`
+        ```{r overview1}
+        dgp <- create_dgp(
+            function(n = 100, p = 10, noise_max = 5) {
+            noise_level <- runif(1, max = noise_max)
+            X <- matrix(rnorm(n*p), nrow = n)
+            y <- X %*% rnorm(p, sd = 2) + rnorm(n, sd = noise_level)
+            # dgp's should return a named list
+            return(list(X = X, y = y, noise_level = noise_level))
+            }
+        )
+        dgp
+        ```
+    b. Create Method(s) via `create_method()`
+        ```{r overview2}
+        method <- create_method(
+          # method function args include the names in the dgp output list (`X` and `y` here)
+          function(X, y, ...) {
+            # by concatenating ... with the list output, we can pass through other
+            # outputs from the dgp to later stages of the simulation (evaluation and
+            # visualization)
+            return(c(list(fit = lm(y ~ X)), ...))
+          }
+        )
+        method
+        ```
+    c. Create Evaluator(s) via `create_evaluator()`
+        ```{r overview3}
+        evaluator <- create_evaluator(
+          # the main computational loop of the simulation will return a `tibble::tibble`
+          # which is passed as 'fit_results' to our evaluation functions
+          function(fit_results) {
+            # calculate R-squared
+            fit_results %>%
+              dplyr::mutate(
+                rsq = sapply(fit, function(.fit) summary(.fit)$r.squared)
+              )
+          }
+        )
+        evaluator
+        ```
+    d. Create Visualizer(s) via `create_visualizer()`
+        ```{r overview4}
+        visualizer <- create_visualizer(
+          # you can use 'fit_results' as well here by adding it as an argument
+          function(eval_results) {
+            require(ggplot2)
+            # return a plot r-squared vs noise level
+            ggplot(aes(x = noise_level, y = rsq), data = eval_results[[1]]) +
+              geom_point() +
+              geom_smooth()
+          }
+        )
+        visualizer
+        ```
+2. Assemble these recipe parts into a complete simulation experiment, e.g.:
+    ```{r overview5}
[…]
