+++ vignettes/simChef.Rmd
@@ -4,7 +4,7 @@ output:
   bookdown::html_document2:
     base_format: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{simChef Vignette}
+  %\VignetteIndexEntry{Getting started with simChef}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -39,26 +39,26 @@ grammar](https://design.tidyverse.org/). The core features of `simChef` include:
 1. A tidyverse-inspired grammar of data-driven simulation experiments
 2. A growing built-in library of composable building blocks for evaluation metrics and visualization utilities
 3. Flexible and seamless integration with distributed computation, caching, and checkpointing
-4. Automated generation of an R Markdown document to easily navigate, visualize, and interpret simulation results (example [here](../linear_regression_output.html))
+4. Automated generation of an R Markdown document to easily navigate, visualize, and interpret simulation results (see example [here](../example_experiment.html))
+
+To highlight the ease of running simulation studies using `simChef`, we will begin by showcasing a basic example usage in Section \@ref(example-usage). Then, in Section \@ref(writing-your-own-simulation-experiment), we will provide additional details on how to create your own simulation experiment using `simChef`.
 
 # Example usage
-
-To highlight the ease of running simulation studies using `simChef`, we begin with a basic example usage and defer details to Section \@ref(writing-your-own-simulation-experiment).
 
 > **Simulation Example:** Suppose, for concreteness, that we would like to evaluate the prediction accuracy of two methods (linear regression and random forests) under a linear data-generating process across different signal-to-noise ratios.
 
 At its core, `simChef` breaks down a simulation experiment into four components:
 
 - `DGP`: the data-generating processes from which to *generate* data
-- `Method`: the methods (or models) to *fit* in the experiment
+- `Method`: the methods (or models) to *fit* on the data in the experiment
 - `Evaluator`: the evaluation metrics used to *evaluate* the methods' performance
 - `Visualizer`: the visualization functions used to *visualize* outputs from the method fits or evaluation results (can be tables, plots, or even R Markdown snippets to display)
 
 Using these components, users can easily run a simulation experiment in six simple steps. Below, we summarize these steps and provide code for our running example.
 
-### Step 1. Specify DGP, method, evaluation, and visualization functions of interest. {-}
+### Step 1. Specify/write DGP, method, evaluation, and visualization functions of interest. {-}
 
-Recall that in the given simulation example, we would like to study two methods (linear regression and random forests) under a linear DGP. We code this DGP (`linear_dgp_fun`) and methods (`linear_reg_fun` and `rf_fun`) via custom functions below. 
+Recall that in our simulation example, we would like to study two methods (linear regression and random forests) under a linear DGP. We code this DGP (`linear_dgp_fun`) and methods (`linear_reg_fun` and `rf_fun`) via custom functions below. 
 
 ```{r dgp-funs}
 #' Linear Data-Generating Process
@@ -114,13 +114,13 @@ rf_fun <- function(X_train, y_train, X_test, y_test, ...) {
 }
 ```
 
[…]
