+++ vignettes/simChef_2methods.Rmd
@@ -1,250 +1,0 @@
----
-title: "Getting started with simChef"
-output:
-  bookdown::html_document2:
-    base_format: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{simChef Vignette}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  # disable R chunk evaluation when using macOS in GH Actions
-  # see https://github.com/tidyverse/ggplot2/issues/2252#issuecomment-1006713187
-  # and https://github.com/lcolladotor/biocthis/issues/27 and
-  # https://github.com/bodkan/slendr/commit/f0f5ae18452cc9df5d151874e41b0b8fd5f29aa2#
-  eval = Sys.getenv("RUNNER_OS") != "macOS"
-)
-
-library(simChef)
-
-set.seed(12345)
-
-# remove old cached results to start fresh
-for (fname in list.files(file.path("results", "Example Experiment"),
-                         pattern = ".rds", recursive = T, full.names = T)) {
-  file.remove(fname)
-}
-```
-
-# Overview
-
-Welcome to `simChef`! Our goal is to empower data scientists to focus their attention toward scientific best practices by removing the administrative burdens of simulation design. Using `simChef`, practitioners can seamlessly and efficiently run simulation experiments using an intuitive [tidy
-grammar](https://design.tidyverse.org/). The core features of `simChef` include:
-
-1. A tidyverse-inspired grammar of data-driven simulation experiments
-2. A growing built-in library of composable building blocks for evaluation metrics and visualization utilities
-3. Flexible and seamless integration with distributed computation, caching, and checkpointing
-4. Automated generation of an R Markdown document to easily navigate, visualize, and interpret simulation results (example [here](../linear_regression_output.html))
-
-# Example usage
-
-To highlight the ease of running simulation studies using `simChef`, we begin with a basic example usage and defer details to Section \@ref(writing-your-own-simulation-experiment).
-
-> **Simulation Example:** Suppose, for concreteness, that we would like to evaluate the prediction accuracy of two methods (linear regression and random forests) under two different underlying data-generating processes (a linear and non-linear response function) and across different signal-to-noise ratios.
-
-At its core, `simChef` breaks down a simulation experiment into four components:
-
-- `DGP`: the data-generating processes from which to *generate* data
-- `Method`: the methods (or models) to *fit* in the experiment
-- `Evaluator`: the evaluation metrics used to *evaluate* the methods' performance
-- `Visualizer`: the visualization functions used to *visualize* outputs from the method fits or evaluation results (can be tables, plots, or even R Markdown snippets to display)
-
-Using these components, users can easily run a simulation experiment in six simple steps. Below, we summarize these steps and provide code for our running example.
-
-### Step 1. Specify DGP, method, evaluation, and visualization functions of interest. {-}
-
-Recall that in the given simulation example, we would like to study two methods (linear regression and random forests) under two different DGPs (a linear and non-linear response function). We code these DGPs (`linear_dgp_fun` and `nonlinear_dgp_fun`) and methods (`linear_reg_fun` and `rf_fun`) via custom functions below. 
-
-```{r dgp-funs}
-#' Linear Data-Generating Process
-#' 
-#' @description Generate training and test data according to the classical
[…]
