### 2023-08-21 0bf0e64
minor updates to intro simChef vignette
