+++ vignettes/simChef-full.Rmd
@@ -1,17 +1,20 @@
 ---
 title: "simChef: A comprehensive guide"
 output:
-  rmarkdown::html_vignette
+  bookdown::html_document2:
+    base_format: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{simChef: A comprehensive guide}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+bibliography: references.bib
 ---
 
 ```{r include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
+  fig.align = "center",
   # disable R chunk evaluation when using macOS in GH Actions
   # see https://github.com/tidyverse/ggplot2/issues/2252#issuecomment-1006713187
   # and https://github.com/lcolladotor/biocthis/issues/27 and
@@ -24,115 +27,164 @@ library(simChef)
 set.seed(12345)
 
 # remove old cached results to start fresh
-for (fname in list.files(file.path("results", "Linear Regression Experiment"),
-                         pattern = ".rds", recursive = T, full.names = T)) {
-  file.remove(fname)
+exp_names <- c(
+  "Linear Regression Experiment", "checkpoint-exper", "debugging-exper"
+)
+for (exp_name in exp_names) {
+  fnames <- list.files(
+    file.path("results", exp_name), pattern = ".rds", 
+    recursive = TRUE, full.names = TRUE
+  )
+  for (fname in fnames) {
+    file.remove(fname)
+  }
 }
 ```
 
-# Overview
+```{css echo = FALSE}
+.infobox {
+  padding: .5em 1em .5em 1.5em;
+  border: 2px solid orange;
+  border-radius: 10px;
+  background: #f5f5f5 5px center/3em no-repeat;
+}
+
+.infobox-blue {
+  padding: .5em 1em .5em 1.5em;
+  border: 2px solid cornflowerblue;
+  border-radius: 10px;
+  background: #f5f5f5 5px center/3em no-repeat;
+}
+```
+
+
+# Overview {-}
 
 The goal of `simChef` is to seamlessly and efficiently run simulation experiments using a simple grammar. The results of these simulation experiments can also be conveniently viewed in an organized and interactive browser (or html file) (e.g., [here](../linear_regression_output.html)). The basic usage of `simChef` can be summarized as follows:
 
-1. Create individual parts of the simulation experiment recipe
-    a. Create DGP(s) (data-generating processes) via `create_dgp()`
-        ```{r overview1}
-        dgp <- create_dgp(
-            function(n = 100, p = 10, noise_max = 5) {
-            noise_level <- runif(1, max = noise_max)
-            X <- matrix(rnorm(n*p), nrow = n)
-            y <- X %*% rnorm(p, sd = 2) + rnorm(n, sd = noise_level)
-            # dgp's should return a named list
-            return(list(X = X, y = y, noise_level = noise_level))
-            }
+1. **Define data-generation, method, evaluation, and visualization functions of interest.**
+    ```{r overview-define, eval = FALSE}
+    ## Define data-generating process (DGP) function(s)
+    dgp_fun <- function(n = 100, p = 10, noise_sd = 1) {
+      X <- matrix(rnorm(n*p), nrow = n)
+      betas <- rnorm(p, sd = 2)
+      noise <- rnorm(n, sd = noise_sd)
+      y <- X %*% betas + noise
+      # dgp function should return a named list
+      return(list(X = X, y = y, noise_sd = noise_sd))
+    }
+    
+    ## Define method function(s)
+    # method function args include the names in the dgp output list (`X` and `y` here)
+    method_fun <- function(X, y, ...) {
+      fit <- lm(y ~ X)
[…]
