### 2023-08-21 64f287c
add vignette for inputs/outputs cheat sheet
