### 2023-08-21 b5f4227
restructure docs folder to accomodate multiple example experiments
