+++ vignettes/example-docs/Linear Regression Experiment/docs/objectives.md
@@ -1,0 +1,5 @@
+The objective of this simulation experiment is to provide a toy example on how to use `simChef` and showcase the automated R Markdown-generated documentation. For the sake of illustration, this toy simulation experiment studies the performance of linear regression at the surface-level with the sole purpose of facilitating an easy-to-understand walkthrough.
+
+<span style="color: blue">
+	[Typically, the objective of the simulation experiment (and this blurb) will be more scientific than instructive and will warrant additional context/background and domain knowledge.]
+</span>
