### 2023-08-21 371a6f4
minor link and formatting fixes in vignettes
