+++ vignettes/simChef.Rmd
@@ -3,10 +3,13 @@ title: "Getting started with simChef"
 output:
   bookdown::html_document2:
     base_format: rmarkdown::html_vignette
+    toc: true
 vignette: >
   %\VignetteIndexEntry{Getting started with simChef}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+pkgdown: 
+  as_is: true
 ---
 
 ```{r include = FALSE}
@@ -49,14 +52,14 @@ To highlight the ease of running simulation studies using `simChef`, we will beg
 
 At its core, `simChef` breaks down a simulation experiment into four components:
 
-- `DGP`: the data-generating processes (DGPs) from which to *generate* data
-- `Method`: the methods (or models) to *fit* on the data in the experiment
-- `Evaluator`: the evaluation metrics used to *evaluate* the methods' performance
-- `Visualizer`: the visualization procedures used to *visualize* outputs from the method fits or evaluation results (can be tables, plots, or even R Markdown snippets to display)
+- `DGP()`: the data-generating processes (DGPs) from which to *generate* data
+- `Method()`: the methods (or models) to *fit* on the data in the experiment
+- `Evaluator()`: the evaluation metrics used to *evaluate* the methods' performance
+- `Visualizer()`: the visualization procedures used to *visualize* outputs from the method fits or evaluation results (can be tables, plots, or even R Markdown snippets to display)
 
 Using these components, users can easily run a simulation experiment in six simple steps. Below, we summarize these steps and provide code for our running example.
 
-### Step 1. Define DGP, method, evaluation, and visualization functions of interest. {-}
+## Step 1. Define DGP, method, evaluation, and visualization functions of interest. {-}
 
 Recall that in our simulation example, we would like to study two methods (linear regression and random forests) under a linear DGP. We code this DGP (`linear_dgp_fun`) and methods (`linear_reg_fun` and `rf_fun`) via custom functions below. 
 
@@ -118,9 +121,9 @@ We can similarly write custom functions to evaluate the prediction accuracy of t
 
 *Note*: Most of the user-written code is encapsulated by Step 1. From here, there is minimal coding on the user end as we turn to leverage the `simChef` grammar for experiments.
 
-### Step 2. Convert functions into `DGP`, `Method`, `Evaluator`, and `Visualizer` class objects. {-}
+## Step 2. Convert functions into `DGP()`, `Method()`, `Evaluator()`, and `Visualizer()` class objects. {-}
 
-Once we have specified the relevant DGP, method, evaluation, and visualization functions, the next step is to convert these functions into `DGP`, `Method`, `Evaluator`, and `Visualizer` class objects (`R6`) for `simChef`. To do so, we simply wrap the functions in `create_dgp()`, `create_method()`, `create_evaluator()`, or `create_visualizer()`, while specifying (a) an intelligible name for the object and (b) any input parameters to pass to the function. For example, 
+Once we have specified the relevant DGP, method, evaluation, and visualization functions, the next step is to convert these functions into `DGP()`, `Method()`, `Evaluator()`, and `Visualizer()` class objects (`R6`) for `simChef`. To do so, we simply wrap the functions in `create_dgp()`, `create_method()`, `create_evaluator()`, or `create_visualizer()`, while specifying (a) an intelligible name for the object and (b) any input parameters to pass to the function. For example, 
 
 ```{r create-classes}
 ## DGPs
@@ -156,7 +159,7 @@ pred_err_plot <- create_visualizer(
 ) 
 ```
 
-### Step 3. Assemble recipe parts into a complete simulation experiment. {-}
+## Step 3. Assemble recipe parts into a complete simulation experiment. {-}
 
[…]
