+++ vignettes/cheat-sheet.Rmd
@@ -24,10 +24,7 @@ library(simChef)
 set.seed(12345)
 ```
 
-```{r fit-eval-viz-figure, fig.cap = "Overview of inputs and outputs for user-defined functions when running a `simChef` `Experiment`. Dotted boxes denote optional input arguments that are automatically passed through in a `run_experiment()` call if specified. ", out.width = "100%", echo = FALSE}
-knitr::include_graphics(path = "figures/fit_eval_viz.png")
-```
-
+![Overview of inputs and outputs for user-defined functions when running a `simChef` `Experiment`. Dotted boxes denote optional input arguments that are automatically passed through in a `run_experiment()` call if specified.](figures/fit_eval_viz.png){width=100%" #fig:fit-eval-viz-figure}
 
 ### DGP 
 
