+++ vignettes/simChef-full.Rmd
@@ -3,10 +3,13 @@ title: "simChef: A comprehensive guide"
 output:
   bookdown::html_document2:
     base_format: rmarkdown::html_vignette
+    toc: true
 vignette: >
   %\VignetteIndexEntry{simChef: A comprehensive guide}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
+pkgdown: 
+  as_is: true
 bibliography: references.bib
 ---
 
@@ -43,17 +46,23 @@ for (exp_name in exp_names) {
 
 ```{css echo = FALSE}
 .infobox {
-  padding: .5em 1em .5em 1.5em;
+  padding: .5em 1em 0em 1.5em;
   border: 2px solid orange;
   border-radius: 10px;
   background: #f5f5f5 5px center/3em no-repeat;
+  margin-bottom: 1em;
 }
 
 .infobox-blue {
-  padding: .5em 1em .5em 1.5em;
+  padding: .5em 1em 0em 1.5em;
   border: 2px solid cornflowerblue;
   border-radius: 10px;
   background: #f5f5f5 5px center/3em no-repeat;
+  margin-bottom: 1em;
+}
+
+details {
+  margin-bottom: unset
 }
 ```
 
@@ -104,7 +113,7 @@ The goal of `simChef` is to seamlessly and efficiently run simulation experiment
         geom_smooth()
     }
     ```
-2. **Convert functions into `DGP`, `Method`, `Evaluator`, and `Visualizer` class objects.**
+2. **Convert functions into `DGP()`, `Method()`, `Evaluator()`, and `Visualizer()` class objects.**
     ```{r overview-class, eval = FALSE}
     dgp <- create_dgp(.dgp_fun = dgp_fun, .name = "DGP1")
     method <- create_method(.method_fun = method_fun, .name = "Method1")
@@ -140,7 +149,7 @@ In this vignette, we will provide an overview of the core features of `simChef`:
 3. Flexible and seamless integration with distributed computation, caching, checkpointing, and debugging tools
 4. Automated generation of an R Markdown document to easily navigate, visualize, and interpret simulation results (see example [here](../linear_regression_output.html))
     
-Note: Before proceeding through this guide, we recommend first checking out our introductory vignette: [Getting started with simChef](`simChef.html`).
+Note: Before proceeding through this guide, we recommend first checking out our introductory vignette: [Getting started with simChef](simChef.html).
 
 # A tidy grammar for simulation experiments
 
@@ -168,7 +177,7 @@ There is also a fifth and final noun:
 Leveraging this grammar, we can easily use `simChef` to run a simulation experiment in six steps.
 
 1. Define DGP, method, evaluation, and visualization functions of interest.
-2. Convert functions into `DGP`, `Method`, `Evaluator`, and `Visualizer` class objects.
+2. Convert functions into `DGP()`, `Method()`, `Evaluator()`, and `Visualizer()` class objects.
 3. Assemble objects into a complete simulation experiment.
 4. Document and describe the simulation experiment in text.
 5. Run the experiment.
@@ -180,11 +189,11 @@ Leveraging this grammar, we can easily use `simChef` to run a simulation experim
 
 To begin, we first need to define the individual parts of the simulation experiment, namely, the data-generation, method, evaluation, and visualization functions under study. 
 We also need to write these functions such that they "play nicely" with one another under the `simChef` API.
-We will make note of these requirements as we go, but for an introductory overview, please refer to the *Writing your own simulation experiment* section in [Getting started with simChef](`simChef.html`) or our [Cheat sheet: Inputs/outputs for user-defined functions](`cheat-sheet.html`).
[…]
