### 2023-08-22 #5642 @dhhagan
@rcannood - I think I overlooked your earlier comment on the expectations of reviewers and authors. Our guidelines for authors can be found [here](https://joss.readthedocs.io/en/latest/submitting.html) and for reviewers [here](https://joss.readthedocs.io/en/latest/review_criteria.html). If you have additional questions, I'd be happy to answer.
