### 2023-09-22 9d58deb
Vignette updates (#170)
* create intro simChef vignette

* update comprehensive simChef vignette

* add vignette for inputs/outputs cheat sheet

* restructure docs folder to accomodate multiple example experiments

Close #160
