+++ vignettes/simChef-full.Rmd
@@ -1,0 +1,1260 @@
+---
+title: "simChef: A comprehensive guide"
+output:
+  bookdown::html_document2:
+    base_format: rmarkdown::html_vignette
+    toc: true
+vignette: >
+  %\VignetteIndexEntry{simChef: A comprehensive guide}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+pkgdown: 
+  as_is: true
+bibliography: references.bib
+---
+
+```{r include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.align = "center",
+  # disable R chunk evaluation when using macOS in GH Actions
+  # see https://github.com/tidyverse/ggplot2/issues/2252#issuecomment-1006713187
+  # and https://github.com/lcolladotor/biocthis/issues/27 and
+  # https://github.com/bodkan/slendr/commit/f0f5ae18452cc9df5d151874e41b0b8fd5f29aa2#
+  eval = Sys.getenv("RUNNER_OS") != "macOS"
+)
+
+library(simChef)
+
+set.seed(12345)
+
+# remove old cached results to start fresh
+exp_names <- c(
+  "Linear Regression Experiment", "checkpoint-exper", "debugging-exper"
+)
+for (exp_name in exp_names) {
+  fnames <- list.files(
+    file.path("results", exp_name), pattern = ".rds", 
+    recursive = TRUE, full.names = TRUE
+  )
+  for (fname in fnames) {
+    file.remove(fname)
+  }
+}
+```
+
+```{css echo = FALSE}
+.infobox {
+  padding: .5em 1em 0em 1.5em;
+  border: 2px solid orange;
+  border-radius: 10px;
+  background: #f5f5f5 5px center/3em no-repeat;
+  margin-bottom: 1em;
+}
+
+.infobox-blue {
+  padding: .5em 1em 0em 1.5em;
+  border: 2px solid cornflowerblue;
+  border-radius: 10px;
+  background: #f5f5f5 5px center/3em no-repeat;
+  margin-bottom: 1em;
+}
+
+details {
+  margin-bottom: unset
+}
+```
+
+
+# Overview {-}
+
+The goal of `simChef` is to seamlessly and efficiently run simulation experiments using a simple grammar. The results of these simulation experiments can also be conveniently viewed in an organized and interactive browser (or html file) (e.g., [here](../linear_regression_output.html)). The basic usage of `simChef` can be summarized as follows:
+
+1. **Define data-generation, method, evaluation, and visualization functions of interest.**
+    ```{r overview-define, eval = FALSE}
+    ## Define data-generating process (DGP) function(s)
+    dgp_fun <- function(n = 100, p = 10, noise_sd = 1) {
+      X <- matrix(rnorm(n*p), nrow = n)
+      betas <- rnorm(p, sd = 2)
+      noise <- rnorm(n, sd = noise_sd)
+      y <- X %*% betas + noise
+      # dgp function should return a named list
+      return(list(X = X, y = y, noise_sd = noise_sd))
+    }
+    
+    ## Define method function(s)
+    # method function args include the names in the dgp output list (`X` and `y` here)
+    method_fun <- function(X, y, ...) {
+      fit <- lm(y ~ X)
+      # by concatenating ... with the list output, we can easily pass through 
+      # other outputs from the dgp to later stages of the simulation 
+      # (evaluation and visualization)
+      return(list(fit = fit, ...))
+    }
+
+    ## Define evaluation function(s)
+    # the main computational loop of the simulation will return a `tibble::tibble`
+    # which is passed as 'fit_results' to our evaluation functions
[…]
