+++ vignettes/example-docs/Example Experiment/docs/visualizers/Prediction Accuracy Plot.md
@@ -1,0 +1,5 @@
+We plot the prediction accuracy between the true and predicted responses, as measured via RMSE, MAE, and $R^2$, to understand how characteristics of the DGP affect various prediction methods.
+
+<span style="color: blue">
+	[In practice, documentation of the plotters should answer the questions “what” and “why”. That is, “what” is the plot, and “why” are we using/studying it?]
+</span>
