+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   simChef-full: simChef-full.html
   simChef: simChef.html
   simulation-directory-setup: simulation-directory-setup.html
-last_built: 2023-09-22T20:53Z
+last_built: 2023-09-22T21:18Z
 urls:
   reference: https://yu-group.github.io/simChef/reference
   article: https://yu-group.github.io/simChef/articles
