### 2023-09-23 895f434
fix broken figure cross references
