### 2023-12-18 #6156 @editorialbot
[REVIEW]: simChef: High-quality data science simulations in R
**Submitting author:** @jpdunc23 (<a href="http://orcid.org/0000-0003-3297-681X">James Duncan</a>)
**Repository:** https://github.com/Yu-Group/simChef
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.0

@rcannood & @Abinashbunty & @anandojha & @Anshuman5, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
