### 2023-12-18 Yu-Group/simChef#171 @rcannood
[JOSS] Comments on the software paper
This issue contains my review of the Software paper of the JOSS submission of this software at openjournals/joss-reviews#6156. At first glance, the paper already looks very well written. Going through the JOSS criteria, I did notice a few things that could be improved.

* The structure of the manuscript is incorrect (See [docs](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)). For one, there should first be a section called "Summary", which has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience. There is a section called "Summary" which partially contains this information, but does not function as a standalone, high level description of the software (since it depends on the "Statement of need" for context)

* The section called "Statement of need" should come after the Summary. I fully agree with the content that the authors wrote, yet I feel like the phrasing of this section makes it a little hard to read, in that it consists of long sentences with quite complex wording. Perhaps simply splitting up this section into three separate paragraphs would already do the trick.

* A section "State of the field" as required by JOSS is missing.

Thanks for your efforts in reviewing my suggestions! If you resolve any of the items above, please link the PRs is this issue.
