### 2023-12-19 ad70520
Rework 'Summary' and 'Statement of need' sections
