### 2024-01-12 #6156 @rcannood
Thanks for poking us!

For the manuscript I'm currently waiting on the authors to complete Yu-Group/simChef#171. For the software, I asked some extra questions in Yu-Group/simChef#173 to better understand if the framework does what I think it's supposed to do.
