### 2024-01-15 Yu-Group/simChef#174 @rcannood
[JOSS] Comments on the documentation
This issue contains the continuation of my review of the JOSS submission of this tool at https://github.com/openjournals/joss-reviews/issues/6156, this time concerning the Documentation.

## Checklist

- [x] Home page - Fix example in overview?
- [x] Home page - Add figure to concepts
- [x] Home page - Thinking like a chef?
- [x] Example usage - Showcase real world example
- [x] Functionality documentation - Add examples
- [x] Community guidelines - Add guidelines for third parties to contribute to the software

## [Home page](https://yu-group.github.io/simChef/index.html)

### Fix example in overview?

The example shown in the first section "Overview" is not something that users can _actually_ run. Is there a way to make turn this into something users _can_ run, e.g. by using predefined functions for `dgp`, `method`, `evaluator`, `visualizer`?

### Add figure to concepts
I also wonder if a schematic visualisation in the section "Concepts" could be valuable, to visually show how these concepts relate to one another. Something like this figure?

![](https://yu-group.github.io/simChef/articles/figures/run_experiment.png)

### Thinking like a chef?

Maybe it was fun to write and fun for some people to read, but the analogies used in this section make it hard to find out what information is actually relevant w.r.t. simChef. I tend to skip the info about the cooking and just read the last two sentences. Maybe it's just me, though.

## Example usage

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The examples included in the documentation all seem to be toy examples, which made it harder for me to understand at first what the authors meant _exactly_ by all of the different concepts and terminology. At the end of the [Get started](https://yu-group.github.io/simChef/articles/simChef.html) page is a really good use case of the tool, however, which really made things click for me. It's also a nice showcase of the rendered documentation.

## Functionality documentation

> Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Mostly! The reference documentation should include working examples for each of the documented functions. CRAN will require you provide this anyways.

## Community guidelines

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Please write some contribution guidelines
