### 2024-01-15 Yu-Group/simChef#173 @rcannood
I found this [case study](https://github.com/PhilBoileau/simChef-case-study) and the corresponding [documentation](https://philboileau.github.io/simChef-case-study/results/empirical-fdr-comparison/empirical-fdr-comparison.html) in the "[Get started](https://yu-group.github.io/simChef/articles/simChef.html)" page of simChef. This confirms my previous suspicions of what simChef is and does.

I also found a response to my question on the vignette related to [computing experimental replicates in parallel](https://yu-group.github.io/simChef/articles/parallel.html). I'll create a separate issue to follow up on this.

Closing this issue since I found answers to my own questions :)
