### 2024-01-15 Yu-Group/simChef#176 @jpdunc23
`create_sim`: improvements to docs and vignettes
Motivated by #175, the [documentation](https://yu-group.github.io/simChef/articles/simulation-directory-setup.html) and [vignette](https://yu-group.github.io/simChef/articles/simulation-directory-setup.html) for `create_sim` could be improved. Here are some specific items:

- [x] Change "character" to "string" in the description of the `path` arg.
- [x] Change "high-power computing" to "high-performance computing" throughout.
- [x] Add examples with file trees for various combinations of the args. Could some or all of the vignette be incorporated into the docs?
- [x] Incorporate `create_sim()` in the [Getting started with simChef](https://yu-group.github.io/simChef/articles/simChef.html) vignette.

More generally, we should emphasize `create_sim()` as an _optional_ first step to setting up a new simulation project and clarify our basic integration with `renv` and `git`.

See https://github.com/Yu-Group/simChef/issues/177 as well.
