### 2024-01-15 Yu-Group/simChef#177 @jpdunc23
`create_sim`: functionality and naming consistency
- **Naming**: We have many `create_*` functions that create `R6` objects like the `Experiment`, `DGP`, and `Method` classes. On the other hand, we used `init_docs()` for setting up the Rmd documentation directory structure. I think renaming `create_sim()` to `init_sim()` would improve our API consistency. Another point of confusion is the `hpc` argument, which could mislead users into thinking that with `hpc = TRUE` their simulation will automatically work in a distributed computing environment (e.g., a Slurm cluster). I think we should rename this argument somehow to make it a bit more clear. Maybe `scripts_dir` (and also rename `tests` to `tests_dir` / add a `logs_dir` arg)? Another option (my preference) is to replace the `hpc` and `tests` args with an `init_dirs` argument that accepts string vectors, default `c("logs", "scripts", "tests")`.
- **Functionality**: Can we have `create_sim()` optionally call `init_docs()` ? E.g., we could add a boolean argument `init_docs`.

Maybe `create_sim()` and `init_docs()` should be combined, somehow?
