### 2024-01-15 Yu-Group/simChef#175 @jpdunc23
@rcannood first off, thank you so much for your careful review and very helpful comments. I'll try to address some of your points below. TL;DR: Please provide some feedback

> I found some information in [Setting Up Your Simulation Study](https://yu-group.github.io/simChef/articles/simulation-directory-setup.html) that suggests you can use hpc = TRUE and init_renv = TRUE to enable distributed computations and reproducibility. However, at this stage I find it hard to argue that the package seamlessly integrates with any of these concepts, since they are not discussed enough in the documentation.

I've added a couple of issues related to this part of the API: #176 and #177. I agree this is a bit unclear as it stands but I think that a few updates to the docs and a small bit of API refactoring could go a long way to addressing your concerns.

One point of confusion is the `hpc` argument which I think should be renamed / removed. Setting `hpc = TRUE` does nothing more than create directories called `scripts` and `logs`. We can definitely make this clearer for users.

> Would it be possible to create separate articles in the documentation to showcase how to set up distributed computation, set up reproducibility, how to use caching & checkpointing, and how to debug runs? Or would you have an alternative solution?

We've definitely found it challenging to document `simChef` in a way that is both approachable and comprehensive, so thank you for taking a close look at the docs and giving us your impressions as a new user. As you've seen, we have a [dedicated vignette on using distributed computation](https://yu-group.github.io/simChef/articles/parallel.html). You've probably also seen the ["comprehensive guide" vignette](https://yu-group.github.io/simChef/articles/simChef-full.html) which is quite long. That vignette includes sections on [caching](https://yu-group.github.io/simChef/articles/simChef-full.html#caching), [checkpointing](https://yu-group.github.io/simChef/articles/simChef-full.html#checkpointing), and [debugging](https://yu-group.github.io/simChef/articles/simChef-full.html#error-debugging).

To clarify, do you think it would be better to created separate, self-contained vignettes for each concept? Another option is to split the comprehensive guide into smaller chunks (e.g. Part 1, Part 2, etc) focused on specific parts of the API. Would that achieve the same goal?

> In the context of reproducibility I would expect the rendered report to contain information on software versions of the used environments.

I think this is a great idea. I've opened https://github.com/Yu-Group/simChef/issues/178.

> W.r.t. caching and debugging tools I'd expect to be able to see the execution time, CPU usage, memory usage and error messages when using an HPC as backend. What happens when one of the executions fails? How can I debug what went wrong during a failed run?

We print the total execution time at the end of `Experiment$fit` / `Experiment$run()` (e.g., see [this example](https://yu-group.github.io/simChef/articles/parallel.html#example)) and optionally collect some basic memory usage stats via `gc()`  when the user sets `options(simChef.debug = TRUE)` (documented [here](https://yu-group.github.io/simChef/articles/parallel.html?q=simChef.debug#dealing-with-task-heterogeneity)). As for displaying CPU usage, I agree this would be a nice feature but I'm not sure it's within scope right now and users can use external tools like `top` or `htop` to check CPU usage as the sim is running. But do you maybe want to expand upon what specifically you'd like to see from `simChef` along these lines?

As I mentioned above, debugging capabilities are documented [here](https://yu-group.github.io/simChef/articles/simChef-full.html#error-debugging). When an error occurs, we capture the error object itself, process ID, output from `gc()`, and the specific `DGP` instance (and `Method` instance, if the replicate got that far) and parameters that were in play when the error happened. This way, users can directly inspect and interact with those objects after the fact. During batch execution (e.g., when running an experiment as part of a Slurm sbatch job), the pattern is to catch and save the `simChef_error`, e.g. as in the above vignette:

```
results <- tryCatch(
  run_experiment(buggy_experiment, n_reps = 2),
  error = identity
)
# check if the results obj has error type and save to disk for later interactive inspection
```

I don't blame you if you missed that example in the "comprehensive guide" vignette given that vignette is a monolith, or maybe that example just isn't clear enough. If the problem is the former, then my inclination is to split that vignette into separate parts so that users can quickly get to the information they need via the docs site header navigation.

Please let us know if you think a different approach would be better or if the existing example needs to be reworked.

> This ties in with the previous comment. The paper mentions efficient usage of computational resources, but some simulation studies will require using an HPC to be able to run the analysis in a decent time frame.

This is possible using the appropriate `future` backend. For example, using `plan(multicore)` would enable a small simulation in parallel on a laptop with 8 cores or a modestly-sized simulation using 64 cores on single cluster node. For larger simulations requiring multi-node computations, the backends provided by [future.batchtools](https://future.batchtools.futureverse.org/) should work. Your comments make me think that it would probably be worthwhile to document such a workflow, so I've added issue https://github.com/Yu-Group/simChef/issues/179.
