### 2024-01-29 #6156 @jpdunc23
@Kevin-Mattheus-Moerman sorry for the delay.

https://github.com/Yu-Group/simChef/issues/171

Partially addressed by https://github.com/Yu-Group/simChef/pull/172. I'll work on adding a "State of the field" section today.

https://github.com/Yu-Group/simChef/issues/174

@tiffanymtang could you send an update on this issue?

https://github.com/Yu-Group/simChef/issues/175

I responded to @rcannood 's very helpful comments and currently waiting for a response. Some of the comments are requests for additional features or API changes which we don't have a lot of capacity to address right now. The TODOs that I extracted from this issues are:

- [] Split our [Comprehensive Guide](https://yu-group.github.io/simChef/articles/simChef-full.html) vignette into more manageable chunks.
- [] A bit of refactoring of the `create_sim()` helper: https://github.com/Yu-Group/simChef/issues/176 and https://github.com/Yu-Group/simChef/issues/177.
- [] Document a multi-node workflow: https://github.com/Yu-Group/simChef/issues/179

@rcannood have I missed anything? Of these tasks, which ones are blocking for this JOSS review?
