### 2024-01-29 #6156 @tiffanymtang
Regarding https://github.com/Yu-Group/simChef/issues/174

My apologies for the delay! I am working on this issue and will hopefully have an update next week. Will keep everyone posted. Thanks!
