+++ vignettes/paper.bib
@@ -77,3 +77,94 @@
   year = {2022},
   note = {https://r6.r-lib.org, https://github.com/r-lib/R6/},
 }
+
+@article{chalmers-simdesign-2020,
+    author =    {Chalmers, R. Philip AND Adkins, Mark C. },
+    journal =   {The Quantitative Methods for Psychology},
+    publisher = {TQMP},
+    title =     {Writing Effective and Reliable Monte Carlo Simulations with the SimDesign Package},
+    year =      {2020},
+    volume =    {16},
+    number =    {4},
+    url =       {http://www.tqmp.org/RegularArticles/vol16-4/p248/p248.pdf },
+    pages =     {248-280},
+    doi =       {10.20982/tqmp.16.4.p248}
+}
+
+@article{gasparini-rsimsum-2018,
+    doi = {10.21105/joss.00739},
+	url = {https://doi.org/10.21105/joss.00739},
+	year = {2018}, publisher = {The Open Journal},
+	volume = {3},
+	number = {26},
+	pages = {739},
+	author = {Gasparini, Alessandro},
+	title = {rsimsum: Summarise results from Monte Carlo simulation studies},
+	journal = {Journal of Open Source Software}
+}
+
+@article{blair-declaredesign-2019,
+    Author = {Blair, Graeme and Cooper, Jasper and Coppock, Alexander and Humphreys, Macartan},
+    Title = {{Declaring and Diagnosing Research Designs}},
+    Journal = {American Political Science Review},
+    Year = {2019},
+    Volume = {113},
+    Number = {3},
+    Pages = {838--859},
+}
+
+@Manual{joshi-simhelpers-2024,
+  title = {simhelpers: Helper Functions for Simulation Studies},
+  author = {Joshi, Megha and Pustejovsky, James},
+  year = {2024},
+  note = {R package version 0.2.0},
+  url = {https://meghapsimatrix.github.io/simhelpers/index.html},
+}
+
+@Manual{scheer-simTool-2020,
+  title = {simTool: Conduct Simulation Studies with a Minimal Amount of Source Code},
+  author = {Scheer, Marsel},
+  year = {2020},
+  note = {R package version 1.1.7},
+  url = {https://cran.r-project.org/web/packages/simTool/index.html},
+}
+
+@Manual{epskamp-parSim-2024,
+  title = {parSim: Parallel Simulation Studies},
+  author = {Sacha Epskamp},
+  year = {2023},
+  note = {R package version 0.1.5},
+  url = {https://cran.r-project.org/web/packages/parSim/parSim.pdf},
+}
+
+@misc{bien-simulator-2016,
+      title={The Simulator: An Engine to Streamline Simulations},
+      author={Jacob Bien},
+      year={2016},
+      eprint={1607.00021},
+      archivePrefix={arXiv},
+      primaryClass={stat.CO}
+}
+
+@Article{couch-infer-2021,
+  title = {{infer}: An {R} package for tidyverse-friendly statistical inference},
+  author = {Simon P. Couch and Andrew P. Bray and Chester Ismay and Evgeni Chasnovski and Benjamin S. Baumer and Mine Çetinkaya-Rundel},
+  journal = {Journal of Open Source Software},
+  year = {2021},
+  volume = {6},
+  number = {65},
+  pages = {3661},
+  doi = {10.21105/joss.03661},
+}
+
+@article{hofert-simsalapar-2016,
+ title={Parallel and Other Simulations in R Made Easy: An End-to-End Study},
+ volume={69},
+ url={https://www.jstatsoft.org/index.php/jss/article/view/v069i04},
+ doi={10.18637/jss.v069.i04},
+ number={4},
[…]
