+++ vignettes/paper.md
@@ -41,7 +41,7 @@ bibliography: paper.bib
 
 # Summary
 
-`simChef` is an R package that empowers data science practitioners to rapidly
+`simChef` is an `R` package that empowers data science practitioners to rapidly
 plan, carry out, and summarize statistical simulation studies in a flexible,
 efficient, and low-code manner. Drawing substantially from the Predictability,
 Computability, and Stability (PCS) framework [@yu-veridical-2020], `simChef`
@@ -92,7 +92,7 @@ At its core, `simChef` breaks down a simulation experiment into four modular com
 - `DGP`: the data-generating processes from which to *generate* data
 - `Method`: the methods (or models) to *fit* in the experiment
 - `Evaluator`: the evaluation metrics used to *evaluate* the methods' performance
-- `Visualizer`: the visualization functions used to *visualize* outputs from the method fits or evaluation results (can be tables, plots, or even R Markdown snippets to display)
+- `Visualizer`: the visualization functions used to *visualize* outputs from the method fits or evaluation results (can be tables, plots, or even `R` Markdown snippets to display)
 
 ![Overview of the four core components in a `simChef` `Experiment`. `simChef` 
 provides four classes that implement distinct simulation objects in
@@ -200,12 +200,62 @@ Once saved, the user can add new `DGP` and `Method` objects to the experiment an
 Considering the example above, when we add `new_method` and call `run_experiment` with `use_cached = TRUE`, `simChef` finds that the cached results are missing combinations of `new_method`, existing DGPs, and their associated parameters, giving nine new configurations. 
 Replicates for the new combinations are then appended to the cached results.
 
-`simChef` also provides users with a convenient API to automatically generate an R Markdown document. 
+`simChef` also provides users with a convenient API to automatically generate an `R` Markdown document. 
 This documentation gathers the scientific details, summary tables, and visualizations side-by-side with the user's custom source code and parameters for data-generating processes, statistical methods, evaluation metrics, and plots. 
 A call to `init_docs` generates empty markdown files for the user to populate with their overarching simulation objectives and with descriptions of each of the `DGP`, `Method`, `Evaluator`, and `Visualizer` objects included in the `Experiment`. 
-Finally, a call to `render_docs` prepares the R Markdown document, either for iterative design and analysis of the simulation or to provide a high-quality overview that can be shared easily. 
+Finally, a call to `render_docs` prepares the `R` Markdown document, either for iterative design and analysis of the simulation or to provide a high-quality overview that can be shared easily. 
 We provide an example of the simulation documentation [here](https://philboileau.github.io/simChef-case-study/results/empirical-fdr-comparison/empirical-fdr-comparison.html).
-Corresponding R source code is available on [GitHub](https://github.com/PhilBoileau/simChef-case-study).
+Corresponding `R` source code is available on [GitHub](https://github.com/PhilBoileau/simChef-case-study).
+
+# Related `R` packages
+
+A number of existing `R` packages and projects address needs related `simChef`'s
+functionality. The `batchtools` package [@lang-batchtools-2017] provides
+abstractions for "problems", "algorithms", and "experiments", similar to
+`simChef`'s `DGP`, `Method`, and `Experiment` objects, respectively.
+Additionally, `batchtools` provides a number of utilities for shared-memory and
+distributed memory computations, including for interacting with high-performance
+computing cluster schedulers such as Slurm and Torque. `simChef` is able to
+leverage these utilities for distributed computations via the backends provided
+by the `future.batchtools` package which is part of the `future` ecosystem of
+`R` packages [@bengtsson-unifying-2021]. Whereas `batchtools` is a general tool
+for distributed mapping operations, `simChef` specializes in data science
[…]
