### 2024-01-29 #6156 @jpdunc23
Quick update, I just added PR https://github.com/Yu-Group/simChef/pull/180 which finishes work on issue https://github.com/Yu-Group/simChef/issues/171. Once one of my co-authors reviews and that is merged I'll ping again to see if we can close #171.

@Kevin-Mattheus-Moerman thank you for pinging us.
