### 2024-02-01 d26eca3
Improve example in overview
* Add self-contained, executable example usage in readme
* Addresses #174
