+++ README.md
@@ -3,28 +3,15 @@
 ## Overview
 
 The goal of `simChef` is to help you quickly cook up a fully-realized,
-high-quality, reproducible, and transparently documented simulation study using
-an intuitive and tidy grammar of simulation experiments:
+high-quality, reproducible, and transparently-documented simulation study in
+a flexible, efficient, and low-code manner. `simChef` removes many of
+the administrative burdens of simulation design through: 
 
-```r
-experiment <- create_experiment() %>%
-  add_dgp(dgp1) %>%
-  add_dgp(dgp2) %>%
-  add_method(method1) %>%
-  add_evaluator(evaluator1) %>%
-  add_visualizer(visualizer1) %>%
-  add_vary_across(
-    dgp = dgp1,
-    n = c(100, 1000, 10000)
-  ) %>%
-  add_vary_across(
-    method = method1,
-    lambda = c(0.1, 0.5, 1.0)
-  )
-
-results <- experiment %>%
-  run_experiment(n_reps = 100)
-```
+1. An intuitive [tidy grammar](https://design.tidyverse.org/) of data science simulations
+2. Powerful abstractions for distributed simulation processing backed by `future`
+[@bengtsson-unifying-2021]
+3. Automated generation of interactive [R Markdown](https://rmarkdown.rstudio.com/) simulation 
+documentation, situating results next to the workflows needed to reproduce them. 
 
 ### Installation
 
@@ -34,6 +21,156 @@ GitHub, please use:
 ```r
 devtools::install_github("Yu-Group/simChef")
 ```
+
+### Example Usage
+
+Consider the following toy simulation experiment, where we want to study the 
+prediction accuracy of linear regression and random forests under both linear 
+and non-linear data-generating processes for varying signal-to-noise ratios.
+
+Let us first code up the necessary simulation components, namely, the linear
+and nonlinear (here, an exclusive-or) data-generating processes as well as the
+linear regression and random forest models. To evaluate the methods and visualize 
+the results, one can also write custom code, but we will leverage built-in evaluation 
+and visualization functions (e.g., `summarize_pred_err` and `plot_pred_err`) from 
+`simChef` for convenience.
+
+```r
+# Generate data via linear model
+linear_dgp_fun <- function(n_train, n_test, p, beta, noise_sd) {
+  n <- n_train + n_test
+  X <- matrix(rnorm(n * p), nrow = n, ncol = p)
+  y <- X %*% beta + rnorm(n, sd = noise_sd)
+  data_list <- list(
+    X_train = X[1:n_train, , drop = FALSE],
+    y_train = y[1:n_train],
+    X_test = X[(n_train + 1):n, , drop = FALSE],
+    y_test = y[(n_train + 1):n]
+  )
+  return(data_list)
+}
+
+# Generate data via exclusive-or model
+xor_dgp_fun <- function(n_train, n_test, p, thresh, beta, noise_sd) {
+  n <- n_train + n_test
+  X <- matrix(rnorm(n * p), nrow = n, ncol = p)
+  xor <- (((X[, 1] > thresh) + (X[, 2] > thresh)) == 1)
+  y <- beta * xor + rnorm(n, sd = noise_sd)
+  data_list <- list(
+    X_train = X[1:n_train, , drop = FALSE],
+    y_train = y[1:n_train],
+    X_test = X[(n_train + 1):n, , drop = FALSE],
+    y_test = y[(n_train + 1):n]
+  )
+  return(data_list)
+}
+
+# Fit linear regression model
+linear_reg_fun <- function(X_train, y_train, X_test, y_test) {
+  train_df <- dplyr::bind_cols(data.frame(X_train), y = y_train)
+  fit <- lm(y ~ ., data = train_df)
+  predictions <- predict(fit, data.frame(X_test))
[…]
