### 2024-02-01 34f40c2
Update concepts section of readme
* Simplify/revise description of simChef concepts and parts of experiment
* Add run experiment figure
* Add concrete examples of DGP/methods/evals/viz using example usage
* Addresses #174
