### 2024-02-01 104b917
showcase real world example in readme
* add link to real world case study in readme
* add links to example html outputs for toy sims from simChef vignettes
* addresses #174
