### 2024-02-03 d5bb2f8
update docs
* add examples (#174)
* create wrappers for DGP$generate, Method$fit, Evaluator$evaluate, and Visualizer$visualize methods
* fix bug in add_na_counts
