+++ R/visualizer-lib-utils.R
@@ -1,7 +1,7 @@
 #' Arguments that are shared by multiple \code{Visualizer} library functions.
 #'
 #' @name shared_viz_lib_args
-#' 
+#'
 #' @param eval_name Name of \code{Evaluator} containing results to plot. If
 #'   \code{NULL}, the data used for plotting is computed from scratch via
 #'   \code{eval_fun}.
@@ -11,25 +11,26 @@
 #' @param eval_fun_options List of named arguments to pass to \code{eval_fun}.
 #' @param interactive Logical. If \code{TRUE}, returns interactive \code{plotly}
 #'   plots. If \code{FALSE}, returns static \code{ggplot} plots.
-#' @param curve Either "ROC" or "PR" indicating whether to plot the ROC or 
+#' @param curve Either "ROC" or "PR" indicating whether to plot the ROC or
 #'   Precision-Recall curve.
 #' @param show Character vector with elements being one of "boxplot", "point",
 #'   "line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
 #'   construct.
-#' @param ... Additional arguments to pass to \code{plot_eval_constructor()}.
+#' @param ... Additional arguments to pass to
+#'   \code{\link{plot_eval_constructor}}.
 #'
 #' @keywords internal
 NULL
 
 #' Developer function for plotting results from Evaluator.
-#' 
+#'
 #' @description A helper function for developing new \code{Visualizer} plotting
-#'   functions that plot the summarized evaluation results as a boxplot, 
-#'   scatter plot, line plot, or bar plot with or without 1 SD error 
+#'   functions that plot the summarized evaluation results as a boxplot,
+#'   scatter plot, line plot, or bar plot with or without 1 SD error
 #'   bars/ribbons. This function accepts either (1) the name(s) of the
 #'   \code{Evaluator(s)} to plot (\code{eval_names}) or (2) a tibble containing
 #'   the summarized evaluation results to plot (\code{plot_data}).
-#'   
+#'
 #' @inheritParams shared_experiment_helpers_args
 #' @inheritParams shared_viz_lib_args
 #' @param eval_names (Optional) A character vector or string, specifying the
@@ -48,7 +49,7 @@ NULL
 #' @param x_str (Optional) Name of column in data frame to plot on the
 #'   x-axis. Default "auto" chooses what to plot on the x-axis automatically.
 #' @param y_str (Optional) Name of column in data frame to plot on the
-#'   y-axis if \code{show} is anything but "boxplot". Default "auto" chooses 
+#'   y-axis if \code{show} is anything but "boxplot". Default "auto" chooses
 #'   what to plot on the y-axis automatically.
 #' @param y_boxplot_str (Optional) Name of column in data frame to plot on
 #'   the y-axis if \code{show} is "boxplot". Default "auto" chooses
@@ -58,12 +59,12 @@ NULL
 #'   ribbon ggplot layers. Default "auto" chooses what column to use for the
 #'   standard deviations automatically.
 #' @param color_str (Optional) Name of column in data frame to use for the
-#'   color and fill aesthetics when plotting. Default "auto" chooses what to 
+#'   color and fill aesthetics when plotting. Default "auto" chooses what to
 #'   use for the color and fill aesthetics automatically. Use \code{NULL} to
 #'   avoid adding any color and fill aesthetic.
 #' @param linetype_str (Optional) Name of column in data frame to use for
 #'   the linetype aesthetic when plotting. Used only when \code{show = "line"}.
-#'   Default "auto" chooses what to use for the linetype aesthetic 
[…]
