+++ R/evaluator-lib-feature-selection.R
@@ -1,34 +1,34 @@
 #' Evaluate and/or summarize feature selection errors.
-#' 
+#'
 #' @name eval_feature_selection_err_funs
-#' @description Evaluate various feature selection metrics, given the true 
+#' @description Evaluate various feature selection metrics, given the true
 #'   feature support and the estimated feature support.
 #'   \code{eval_feature_selection_err()} evaluates the various feature selection
 #'   metrics for each experimental replicate separately.
 #'   \code{summarize_feature_selection_err()} summarizes the various feature
 #'   selection metrics across experimental replicates.
-#'   
+#'
 #' @inheritParams shared_experiment_helpers_args
 #' @inheritParams shared_eval_lib_args
-#' @param truth_col A character string identifying the column in 
+#' @param truth_col A character string identifying the column in
 #'   \code{fit_results} with the true feature support data. Each element in this
-#'   column should be an array of length \code{p}, where \code{p} is the number 
+#'   column should be an array of length \code{p}, where \code{p} is the number
 #'   of features. Elements in this array should be binary with \code{TRUE} or
-#'   \code{1} meaning the feature (corresponding to that slot) is in the support 
+#'   \code{1} meaning the feature (corresponding to that slot) is in the support
 #'   and \code{FALSE} or \code{0} meaning the feature is not in the support.
 #' @param estimate_col An (optional) character string identifying the column in
 #'   \code{fit_results} with the estimated feature support data. Each element in
-#'   this column should be an array of length \code{p}, where \code{p} is the 
-#'   number of features and the feature order aligns with that of 
-#'   \code{truth_col}. Elements in this array should be binary with \code{TRUE} 
-#'   or \code{1} meaning the feature (corresponding to that slot) is in the 
-#'   estimated support and \code{FALSE} or \code{0} meaning the feature is not 
-#'   in the estimated support. If \code{NULL} (default), the non-zero elements 
+#'   this column should be an array of length \code{p}, where \code{p} is the
+#'   number of features and the feature order aligns with that of
+#'   \code{truth_col}. Elements in this array should be binary with \code{TRUE}
+#'   or \code{1} meaning the feature (corresponding to that slot) is in the
+#'   estimated support and \code{FALSE} or \code{0} meaning the feature is not
+#'   in the estimated support. If \code{NULL} (default), the non-zero elements
 #'   of \code{imp_col} are used as the estimated feature support.
-#' @param imp_col A character string identifying the column in 
-#'   \code{fit_results} with the estimated feature importance data. Each element 
-#'   in this column should be an array of length \code{p}, where \code{p} is the 
-#'   number of features and the feature order aligns with that of 
+#' @param imp_col A character string identifying the column in
+#'   \code{fit_results} with the estimated feature importance data. Each element
+#'   in this column should be an array of length \code{p}, where \code{p} is the
+#'   number of features and the feature order aligns with that of
 #'   \code{truth_col}. Elements in this array should be numeric where a higher
 #'   magnitude indicates a more important feature.
 #' @param metrics A \code{metric_set} object indicating the metrics to evaluate.
@@ -36,12 +36,12 @@
 #'   evaluate the following: number of true positives (\code{tp}), number of
 #'   false positives (\code{fp}), sensitivity (\code{sens}), specificity
 #'   (\code{spec}), positive predictive value (\code{ppv}), number of features
[…]
