+++ man-roxygen/dgp-template.R
@@ -10,4 +10,5 @@
 #'
 #' \itemize{
 #' \item{\code{DGP$new()}} See \link{create_dgp}.
+#' \item{\code{DGP$generate()}} See \link{generate_dgp}.
 #' }
