+++ R/run-tests.R
@@ -4,6 +4,11 @@
 #'   This function only works if the simulation study is set up as an R project
 #'   and if it is run when this R project is active.
 #'
+#' @examples
+#' \dontrun{
+#' # run tests in the tests/ directory
+#' run_tests()}
+#'
 #' @export
 run_tests <- function() {
 
@@ -34,6 +39,11 @@ run_tests <- function() {
 #'   must be stored in sub-directories whose names match those suggesting in the
 #'   "Setting Up Your Simulation" vignette. These details are conveniently taken
 #'   care of when a simulation study is set up using \code{\link{create_sim}()}.
+#'
+#' @examples
+#' \dontrun{
+#' # run dgp-, method-, evaluator-, and visualizer-related tests
+#' test_sim_dir()}
 #'
 #' @export
 test_sim_dir <- function() {
@@ -105,6 +115,11 @@ test_sim_dir <- function() {
 #'   \code{\link[devtools:load_all]{devtools::load_all}()} it loads all of
 #'   the functions in the \code{R/dgp}, \code{R/method}, \code{R/eval} and
 #'   \code{R/viz} directories of the current simulation study.
+#'
+#' @examples
+#' \dontrun{
+#' # load all functions in R/dgp, R/method, R/eval, and R/viz directories
+#' load_all()}
 #'
 #' @export
 load_all <- function() {

[…]
