### 2024-02-03 22134de
reorganize origins of simchef in readme
* move text to collapsible blocks
* bold simChef highlights in "thinking like a chef" to address #174
