+++ README.md
@@ -8,8 +8,7 @@ a flexible, efficient, and low-code manner. `simChef` removes many of
 the administrative burdens of simulation design through: 
 
 1. An intuitive [tidy grammar](https://design.tidyverse.org/) of data science simulations
-2. Powerful abstractions for distributed simulation processing backed by `future`
-[@bengtsson-unifying-2021]
+2. Powerful abstractions for distributed simulation processing backed by [`future`](https://future.futureverse.org/)
 3. Automated generation of interactive [R Markdown](https://rmarkdown.rstudio.com/) simulation 
 documentation, situating results next to the workflows needed to reproduce them. 
 
@@ -170,7 +169,7 @@ For a more detailed walkthrough of this example usage, please see `vignette("sim
 methodology, please check out 
 [Boileau et al.](https://github.com/PhilBoileau/simChef-case-study).**
 The corresponding interactive R Markdown documentation can be found 
-[here](https://philboileau.github.io/simChef-case-study/results/empirical-fdr-comparison/empirical-fdr-comparison.html)
+[here](https://philboileau.github.io/simChef-case-study/results/empirical-fdr-comparison/empirical-fdr-comparison.html).
 
 More examples of the rendered documentation for different simulation experiments:
 - [Toy Example 1](https://yu-group.github.io/simChef/example_experiment.html)
@@ -178,7 +177,7 @@ More examples of the rendered documentation for different simulation experiments
 
 ### Grammar of a `simChef` Simulation Experiment
 
-In `simChef`, a simulation experiment is broken down into four modular concepts, 
+The `simChef` API distills a simulation experiment into four modular concepts, 
 two of which are optional (but highly recommended):
 **data-generating processes** (DGPs), **methods**, **evaluation** (optional),
 and **visualization** (optional). 
@@ -194,7 +193,7 @@ concrete. These four classes are:
     [`dgpoix`](https://yu-group.github.io/dgpoix) (currently in early development).
   - *Ex:* In the above example usage, there are two DGPs: the linear DGP and the 
     exclusive-or DGP.
-- `Method`: corresponds to the **method** (or models) to *fit* to the data in the
+- `Method`: corresponds to the **method** (or model) to *fit* to the data in the
   experiment.
   - Methods can be either a new method under study, a baseline comparison method,
     or any means by which to transform the simulated data (i.e,. the output of DGP).
@@ -218,7 +217,7 @@ concrete. These four classes are:
     replicates.
 
 A fifth `R6` class and concept, `Experiment`, unites the four concepts above. 
-Here, an `Experiment` is a collection of `DGP`(s), `Method`(s), 
+More precisely, an `Experiment` is a collection of `DGP`(s), `Method`(s), 
 `Evaluator`(s), and `Visualizer`(s), which are thoughtfully composed to answer
 a particular question of interest. An `Experiment` can also include references to
 `DGP` and/or `Method` parameters that should be varied and combined during the 
@@ -234,9 +233,12 @@ running the experiment will
 (2) *evaluate* the experiment according to the given `Evaluator`(s), and 
 (3) *visualize* the experiment according to the given `Visualizer`(s).
 
-![Overview of running a `simChef` `Experiment`. The `Experiment` class handles relationships among the four classes: `DGP`, `Method`, `Evaluator`, and `Visualizer`. Experiments may have multiple `DGP`s and `Method`s, which are combined across the Cartesian product of their varying parameters (represented by `\*`). Once computed, each `Evaluator` and `Visualizer` takes in the fitted simulation replicates, while `Visualizer` additionally receives evaluation summaries.](vignettes/figures/run_experiment.png){ width=100% }
[…]
