### 2024-02-03 8853ee8
add contributing guidelines
* address #174
