### 2024-02-06 Yu-Group/simChef#174 @tiffanymtang
@rcannood Thank you so much for your detailed and helpful comments! We really appreciate the time you took to help us improve the `simChef` documentation! I have opened PR #181, pending review, to address your comments here.

> ## [Home page](https://yu-group.github.io/simChef/index.html)
> ### Fix example in overview?
> The example shown in the first section "Overview" is not something that users can _actually_ run. Is there a way to make turn this into something users _can_ run, e.g. by using predefined functions for `dgp`, `method`, `evaluator`, `visualizer`?

Thanks for this great suggestion! We have added a runnable example usage in the readme/homepage. It is admittedly a toy example in order to keep the code as basic and as short as possible.

> ### Add figure to concepts
> I also wonder if a schematic visualisation in the section "Concepts" could be valuable, to visually show how these concepts relate to one another. Something like this figure?
>
> ![](https://camo.githubusercontent.com/087c5cbcb7f6b84a2cce7ad2f25a232cdbaa35a769f210ef28ba904d785b7501/68747470733a2f2f79752d67726f75702e6769746875622e696f2f73696d436865662f61727469636c65732f666967757265732f72756e5f6578706572696d656e742e706e67)

Definitely! We have added this figure and also rewrote the previous "Concepts" section, now called "Grammar of a simChef simulation experiment", to improve its clarity.

> ### Thinking like a chef?
> Maybe it was fun to write and fun for some people to read, but the analogies used in this section make it hard to find out what information is actually relevant w.r.t. simChef. I tend to skip the info about the cooking and just read the last two sentences. Maybe it's just me, though.

We appreciate this comment and agree that it is not the most essential content to understand how to use `simChef`. For this reason, we have de-emphasized this "Origins of `simChef`" section and hid its content in blocks that can be toggled to hide/show. We have also bolded the sentences pertaining to `simChef` features in the "Thinking like a chef" section to help readability in the case one does read this section.

> ## Example usage
> > Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>
> The examples included in the documentation all seem to be toy examples, which made it harder for me to understand at first what the authors meant _exactly_ by all of the different concepts and terminology. At the end of the [Get started](https://yu-group.github.io/simChef/articles/simChef.html) page is a really good use case of the tool, however, which really made things click for me. It's also a nice showcase of the rendered documentation.

Thanks for this feedback! We're really happy to hear that the [Boileau et al.](https://github.com/PhilBoileau/simChef-case-study) case study helped make things click, especially since this was a "true" case study, in that the first author was not involved in the development of `simChef` and was using `simChef` as a first-time user. Since the case study requires some technical statistical background, we prefer not to use this as our primary example usage in the readme/homepage or the "Getting Started" vignette. However, we like the idea of emphasizing it in case it helps understanding, as in your case. To that end, we have included a sentence at the end of the "Example usage" section of the readme/homepage, which highlights this case study and the resulting R Markdown documentation.

> ## Functionality documentation
> > Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> Mostly! The reference documentation should include working examples for each of the documented functions. CRAN will require you provide this anyways.

Thank you for taking the time and care to check on this. We have gone through and added working examples for each of the documented functions.

> ## Community guidelines
> > Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Please write some contribution guidelines

Thanks for pointing this out. We have added contribution guidelines in the new `CONTRIBUTING.md` file.

Thank you again for all of these great suggestions and feedback. Please let us know if there is anything else we can do to improve the documentation or other aspects of `simChef`.
