### 2024-02-06 #6156 @tiffanymtang
Following up on https://github.com/Yu-Group/simChef/issues/174:

I opened PR https://github.com/Yu-Group/simChef/pull/181 to address the comments from @rcannood discussed in https://github.com/Yu-Group/simChef/issues/174. We plan to merge ithese changes soon, pending review from co-authors. Thanks for everyone's diligence!
