+++ man/create_doc_template.Rd
@@ -33,3 +33,4 @@ create_doc_template(experiment)
 create_doc_template(save_dir = experiment$get_save_dir())}
 
 }
+\keyword{internal}
