+++ man/set_rmd_options.Rd
@@ -101,3 +101,4 @@ experiment <- experiment \%>\%
   )
 
 }
+\keyword{internal}
