+++ man/DGP.Rd
@@ -3,27 +3,136 @@
 \docType{class}
 \name{DGP}
 \alias{DGP}
-\title{\code{R6} class representing a data-generating process.}
+\title{\code{R6} class representing a data-generating process}
+\value{
+Result of \code{DGP$dgp_fun()}. If the result is not a list,
+it will be coerced to a list.
+}
 \description{
 A data-generating process which will be used in the
-\code{Experiment} to \strong{generate} data.
+\link{Experiment} to \strong{generate} data.
+}
+\examples{
+# create an example DGP function
+dgp_fun <- function(n, beta, rho, sigma) {
+  cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
+  X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
+  y <- X \%*\% beta + rnorm(n, sd = sigma)
+  return(list(X = X, y = y))
+}
+
+# create DGP (with uncorrelated features)
+dgp <- DGP$new(.dgp_fun = dgp_fun,
+               .name = "Linear Gaussian DGP",
+               # additional named parameters to pass to dgp_fun() by default
+               n = 200, beta = c(1, 0), rho = 0, sigma = 1)
+
+print(dgp)
+
+dgp_uncorr <- dgp$generate()
+cor(dgp_uncorr$X)
+
+dgp_corr <- dgp$generate(rho = 0.7)
+cor(dgp_corr$X)
+
+}
+\seealso{
+\link{create_dgp}
 }
 \section{Public fields}{
-
-
+\if{html}{\out{<div class="r6-fields">}}
 \describe{
 \item{\code{name}}{The name of the \code{DGP}.}
+
 \item{\code{dgp_fun}}{The data-generating process function.}
-\item{\code{dgp_params}}{(Named) list of parameters to input into the data-generating process function.}
+
+\item{\code{dgp_params}}{A (named) list of parameters to input into the data-generating process function.}
 }
+\if{html}{\out{</div>}}
 }
-
-\section{Public methods}{
-
-
+\section{Methods}{
+\subsection{Public methods}{
 \itemize{
-\item{\code{DGP$new()}} See \link{create_dgp}.
-\item{\code{DGP$generate()}} See \link{generate_dgp}.
+\item \href{#method-DGP-new}{\code{DGP$new()}}
+\item \href{#method-DGP-generate}{\code{DGP$generate()}}
+\item \href{#method-DGP-print}{\code{DGP$print()}}
+\item \href{#method-DGP-clone}{\code{DGP$clone()}}
 }
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-DGP-new"></a>}}
+\if{latex}{\out{\hypertarget{method-DGP-new}{}}}
+\subsection{Method \code{new()}}{
+Initialize a new \code{DGP} object.
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{DGP$new(.dgp_fun, .name = NULL, ...)}\if{html}{\out{</div>}}
 }
 
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{.dgp_fun}}{The data-generating process function.}
+
+\item{\code{.name}}{(Optional) An optional name for the \code{DGP}, helpful for later
+identification.}
+
+\item{\code{...}}{Default arguments to pass to \code{.dgp_fun()} when \code{DGP$generate()} is
+called.}
+}
+\if{html}{\out{</div>}}
+}
+}
+\if{html}{\out{<hr>}}
[…]
