+++ NAMESPACE
@@ -43,7 +43,6 @@ export(fit_method)
 export(fp)
 export(fp_vec)
 export(generate_data)
-export(generate_dgp)
 export(get_cached_results)
 export(get_dgps)
 export(get_dot_args)
