+++ DESCRIPTION
@@ -68,7 +68,7 @@ Remotes:
     Yu-Group/vthemes
 Config/testthat/edition: 3
 Encoding: UTF-8
-Roxygen: list(markdown = TRUE, r6 = FALSE)
+Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.2.3
 Collate: 
     'dgp.R'

[…]
