### 2024-02-06 1dbec50
Run `roxygen2md::roxygen2md(scope = simple)`
