+++ R/visualizer-lib-utils.R
@@ -17,7 +17,7 @@
 #'   "line", "bar", "errorbar", "ribbon" indicating what plot layer(s) to
 #'   construct.
 #' @param ... Additional arguments to pass to
-#'   \code{\link{plot_eval_constructor}}.
+#'   [plot_eval_constructor()].
 #'
 #' @keywords internal
 NULL
