+++ man/load_all.Rd
@@ -8,7 +8,7 @@ load_all()
 }
 \description{
 \code{load_all()} is simulation study counterpart to
-\code{\link[devtools:load_all]{devtools::load_all}()} it loads all of
+\code{\link[devtools:load_all]{devtools::load_all()}} it loads all of
 the functions in the \code{R/dgp}, \code{R/method}, \code{R/eval} and
 \code{R/viz} directories of the current simulation study.
 }
