+++ man/create_sim.Rd
@@ -13,8 +13,8 @@ directory.}
 \item{init_git}{A \code{logical} indicating whether to intialize your
 simulation directory as a git repository.}
 
-\item{init_renv}{A \code{logical} stating whether to initialize
-\code{\link[renv:init]{renv}}. Defaults to \code{FALSE}.}
+\item{init_renv}{A \code{logical} stating whether to initialize \code{renv} with
+\code{\link[renv:init]{renv::init()}}. Defaults to \code{FALSE}.}
 
 \item{tests}{A \code{logical} indicating whether to generate sub-directories
 for organizing unit tests. Defaults to \code{TRUE}.}
@@ -25,9 +25,9 @@ organizing files related to high-power computing environments. Defaults to
 }
 \description{
 \code{create_sim()} initializes a directory for your simulation
-study. It wraps around \code{\link[usethis]{create_project}()}, as well as
-\code{\link[usethis:use_git]{usethis::use_git}()} and
-\code{\link[renv:init]{renv::init}()}.
+study. It wraps around \code{\link[usethis:create_package]{usethis::create_project()}}, as well as
+\code{\link[usethis:use_git]{usethis::use_git()}} and
+\code{\link[renv:init]{renv::init()}}.
 }
 \examples{
 \dontrun{
