+++ man/plot_pred_curve.Rd
@@ -43,7 +43,7 @@ Precision-Recall curve.}
 construct.}
 
 \item{...}{Additional arguments to pass to
-\code{\link{plot_eval_constructor}}.}
+\code{\link[=plot_eval_constructor]{plot_eval_constructor()}}.}
 }
 \value{
 If \code{interactive = TRUE}, returns a \code{plotly} object if
