+++ man/shared_viz_lib_args.Rd
@@ -25,7 +25,7 @@ Precision-Recall curve.}
 construct.}
 
 \item{...}{Additional arguments to pass to
-\code{\link{plot_eval_constructor}}.}
+\code{\link[=plot_eval_constructor]{plot_eval_constructor()}}.}
 }
 \description{
 Arguments that are shared by multiple \code{Visualizer} library functions.
