+++ DESCRIPTION
@@ -85,10 +85,10 @@ Collate:
     'globals.R'
     'init-dir.R'
     'method.R'
+    'reexport-magrittr.R'
     'run-tests.R'
     'signals.R'
     'use_templates.R'
-    'utils-pipe.R'
     'utils-rmd.R'
     'utils-templates.R'
     'visualizer-lib-feature-selection.R'
