### 2024-02-07 c3b468a
Update CONTRIBUTING.md
Co-authored-by: James Duncan <jpdunc23@users.noreply.github.com>
