+++ R/method.R
@@ -1,23 +1,171 @@
-#' `R6` class representing a method.
+# NOTE: R6 methods can't use the `@inheritParams` tag. If you update the
+# `@param` tags below then be sure to manually replace the corresponding tags
+# above `Method$initialize()`.
+
+#' Create a new `Method`
+#'
+#' @name create_method
+#'
+#' @description Create a [Method] which can `fit()` data in an [Experiment].
+#'
+#' @param .method_fun The user-defined method function.
+#' @param .name (Optional) The name of the `Method`, helpful for later
+#'   identification.
+#' @param ... User-defined default arguments to pass into `.method_fun()`.
+#'
+#' @return A new [Method] object.
+#'
+#' @examples
+#' # generate some data
+#' dgp_fun <- function(n, beta, rho, sigma) {
+#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
+#'   X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
+#'   y <- X %*% beta + rnorm(n, sd = sigma)
+#'   return(list(X = X, y = y))
+#' }
+#'
+#' dgp <- create_dgp(.dgp_fun = dgp_fun,
+#'                   .name = "Linear Gaussian DGP",
+#'                   n = 50, beta = c(1, 0), rho = 0, sigma = 1)
+#' data_corr <- dgp$generate(rho = 0.7)
+#'
+#' # create an example Method function
+#' lm_fun <- function(X, y, cols) {
+#'   X <- X[, cols]
+#'   lm_fit <- lm(y ~ X)
+#'   pvals <- summary(lm_fit)$coefficients[-1, "Pr(>|t|)"] %>%
+#'     setNames(paste(paste0("X", cols), "p-value"))
+#'   return(pvals)
+#' }
+#'
+#' # create Method with default argument `cols`
+#' lm_method <- create_method(
+#'   .method_fun = lm_fun,
+#'   .name = "OLS",
+#'   cols = c(1, 2)
+#' )
+#'
+#' print(lm_method)
+#'
+#' # fit the Method on data with non-default arguments
+#' lm_method$fit(data_corr, cols = 2)
+#'
+#' # fit the Method on data with default arguments
+#' lm_method$fit(data_corr)
+#'
+#' @export
+create_method <- function(.method_fun, .name = NULL, ...) {
+  Method$new(.method_fun, .name, ...)
+}
+
+#' `R6` class representing a method
+#'
+#' @name Method
 #'
 #' @docType class
 #'
-#' @description A method to be **fit** and assessed in the `Experiment`.
+#' @description `Method` which can `fit()` data in an [Experiment].
 #'
-#' @template method-template
+#'   Generally speaking, users won't directly interact with the `Method` R6
+#'   class, but instead indirectly through [create_method()] and the
+#'   following `Experiment` helpers:
+#'
+#'   - [add_method()]
+#'   - [update_method()]
+#'   - [remove_method()]
+#'   - [get_methods()]
+#'   - [fit_experiment()]
+#'
+#' @seealso [create_method]
+#'
+#' @examples
+#' # generate some data
+#' dgp_fun <- function(n, beta, rho, sigma) {
+#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
+#'   X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
+#'   y <- X %*% beta + rnorm(n, sd = sigma)
+#'   return(list(X = X, y = y))
+#' }
+#'
[…]
