+++ R/dgp.R
@@ -1,15 +1,69 @@
+# NOTE: R6 methods can't use the `@inheritParams` tag. If you update the
+# `@param` tags below then be sure to manually replace the corresponding tags
+# above `DGP$initialize()`.
+
+#' Create a new `DGP` (data-generating process)
+#'
+#' @name create_dgp
+#'
+#' @description Create a [DGP] which can `generate()` data in an [Experiment].
+#'
+#' @param .dgp_fun The user-defined data-generating process function.
+#' @param .name (Optional) An optional name for the `DGP`, helpful for later
+#'   identification.
+#' @param ... User-defined default arguments to pass to `.dgp_fun()` when
+#'   `DGP$generate()` is called.
+#'
+#' @return A new [DGP] object.
+#'
+#' @examples
+#' # create an example DGP function
+#' dgp_fun <- function(n, beta, rho, sigma) {
+#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
+#'   X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
+#'   y <- X %*% beta + rnorm(n, sd = sigma)
+#'   return(list(X = X, y = y))
+#' }
+#'
+#' # create DGP (with uncorrelated features)
+#' dgp <- create_dgp(.dgp_fun = dgp_fun,
+#'                   .name = "Linear Gaussian DGP",
+#'                   # additional named parameters to pass to dgp_fun() by default
+#'                   n = 50, beta = c(1, 0), rho = 0, sigma = 1)
+#'
+#' print(dgp)
+#'
+#' data_uncorr <- dgp$generate()
+#' cor(data_uncorr$X)
+#'
+#' data_corr <- dgp$generate(rho = 0.7)
+#' cor(data_corr$X)
+#'
+#' @export
+create_dgp <- function(.dgp_fun, .name = NULL, ...) {
+  DGP$new(.dgp_fun, .name, ...)
+}
+
 #' `R6` class representing a data-generating process
+#'
+#' @name DGP
 #'
 #' @docType class
 #'
-#' @description A data-generating process which will be used in the
-#'   [Experiment] to **generate** data.
+#' @description `DGP`, a data-generating process which can `generate()` data in
+#'   an [Experiment].
 #'
-#' @param .dgp_fun The data-generating process function.
-#' @param .name (Optional) An optional name for the `DGP`, helpful for later
-#'   identification.
-#' @param ... Default arguments to pass to `.dgp_fun()` when `DGP$generate()` is
-#'   called.
+#'   Generally speaking, users won't directly interact with the `DGP` R6
+#'   class, but instead indirectly through [create_dgp()] and the
+#'   following `Experiment` helpers:
+#'
+#'   - [add_dgp()]
+#'   - [update_dgp()]
+#'   - [remove_dgp()]
+#'   - [get_dgps()]
+#'   - [generate_data()]
+#'
+#' @seealso [create_dgp]
 #'
 #' @examples
 #' # create an example DGP function
@@ -24,32 +78,49 @@
 #' dgp <- DGP$new(.dgp_fun = dgp_fun,
 #'                .name = "Linear Gaussian DGP",
 #'                # additional named parameters to pass to dgp_fun() by default
-#'                n = 200, beta = c(1, 0), rho = 0, sigma = 1)
+#'                n = 50, beta = c(1, 0), rho = 0, sigma = 1)
 #'
 #' print(dgp)
 #'
-#' dgp_uncorr <- dgp$generate()
-#' cor(dgp_uncorr$X)
[…]
