### 2024-02-08 19828d1
Minor doc updates
* improve visualizer example usage
* change T to TRUE in example usage to avoid R CMD CHECK error
* change varying parameter to varying parameter(s)
