+++ R/visualizer.R
@@ -25,77 +25,122 @@
 #' @return A new [Visualizer] object.
 #'
 #' @examples
-#' ## create toy DGPs, Methods, Evaluators, and Visualizers
+#' # create DGP
+#' dgp_fun <- function(n, beta, rho, sigma) {
+#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = TRUE, nrow = 2, ncol = 2)
+#'   X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
+#'   y <- X %*% beta + rnorm(n, sd = sigma)
+#'   return(list(X = X, y = y))
+#' }
+#' dgp <- create_dgp(.dgp_fun = dgp_fun,
+#'                   .name = "Linear Gaussian DGP",
+#'                   n = 50, beta = c(1, 0), rho = 0, sigma = 1)
 #'
-#' # generate toy data
-#' dgp <- create_dgp(
-#'   .dgp_fun = function(n) rnorm(n), .name = "DGP", n = 100
+#' # create Method
+#' lm_fun <- function(X, y, cols) {
+#'   X <- X[, cols]
+#'   lm_fit <- lm(y ~ X)
+#'   pvals <- summary(lm_fit)$coefficients[-1, "Pr(>|t|)"] %>%
+#'     setNames(paste(paste0("X", cols), "p-value"))
+#'   return(pvals)
+#' }
+#' lm_method <- create_method(
+#'   .method_fun = lm_fun,
+#'   .name = "OLS",
+#'   cols = c(1, 2)
 #' )
 #'
-#' # compute mean of data
-#' mean_method <- create_method(
-#'   .method_fun = function(x) list(mean = mean(x)), .name = "Mean(x)"
-#' )
+#' # create an example Evaluator function
+#' reject_prob_fun <- function(fit_results, vary_params = NULL, alpha = 0.05) {
+#'   fit_results[is.na(fit_results)] <- 1
+#'   group_vars <- c(".dgp_name", ".method_name", vary_params)
+#'   eval_out <- fit_results %>%
+#'     dplyr::group_by(across({{group_vars}})) %>%
+#'     dplyr::summarise(
+#'       n_reps = dplyr::n(),
+#'       `X1 Reject Prob.` = mean(`X1 p-value` < alpha),
+#'       `X2 Reject Prob.` = mean(`X2 p-value` < alpha)
+#'     )
+#'   return(eval_out)
+#' }
 #'
-#' # evaluate SD of mean(x) across simulation replicates
-#' sd_mean_eval <- create_evaluator(
-#'   .eval_fun = function(fit_results, vary_params = NULL) {
-#'     group_vars <- c(".dgp_name", ".method_name", vary_params)
-#'     fit_results %>%
-#'       dplyr::group_by(dplyr::across(tidyselect::all_of(group_vars))) %>%
-#'       dplyr::summarise(sd = sd(mean), .groups = "keep")
-#'   },
-#'   .name = "SD of Mean(x)"
-#' )
+#' reject_prob_eval <- Evaluator$new(.eval_fun = reject_prob_fun,
+#'                                   .name = "Rejection Prob (alpha = 0.05)")
 #'
-#' # plot SD of mean(x) across simulation replicates
-#' sd_mean_viz <- create_visualizer(
-#'   .viz_fun = function(fit_results, eval_results, vary_params = NULL,
-#'                       eval_name = "SD of Mean(x)") {
-#'     if (!is.null(vary_params)) {
-#'       add_aes <- ggplot2::aes(
-#'         x = .data[[unique(vary_params)]], y = sd, color = .dgp_name
-#'       )
-#'     } else {
[…]
