+++ R/method.R
@@ -18,7 +18,7 @@
 #' @examples
 #' # generate some data
 #' dgp_fun <- function(n, beta, rho, sigma) {
-#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
+#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = TRUE, nrow = 2, ncol = 2)
 #'   X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
 #'   y <- X %*% beta + rnorm(n, sd = sigma)
 #'   return(list(X = X, y = y))
@@ -81,7 +81,7 @@ create_method <- function(.method_fun, .name = NULL, ...) {
 #' @examples
 #' # generate some data
 #' dgp_fun <- function(n, beta, rho, sigma) {
-#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = T, nrow = 2, ncol = 2)
+#'   cov_mat <- matrix(c(1, rho, rho, 1), byrow = TRUE, nrow = 2, ncol = 2)
 #'   X <- MASS::mvrnorm(n = n, mu = rep(0, 2), Sigma = cov_mat)
 #'   y <- X %*% beta + rnorm(n, sd = sigma)
 #'   return(list(X = X, y = y))
