### 2024-02-08 f40c7c1
update visualizer .Rd docs
