### 2024-02-08 54fb6ce
Expand 'Related R packages' section
