+++ vignettes/paper.bib
@@ -5,7 +5,6 @@
 	issn = {0027-8424, 1091-6490},
 	url = {http://www.pnas.org/lookup/doi/10.1073/pnas.1901326117},
 	doi = {10.1073/pnas.1901326117},
-	abstract = {Building and expanding on principles of statistics, machine learning, and scientific inquiry, we propose the predictability, computability, and stability (PCS) framework for veridical data science. Our framework, composed of both a workflow and documentation, aims to provide responsible, reliable, reproducible, and transparent results across the data science life cycle. The PCS workflow uses predictability as a reality check and considers the importance of computation in data collection/storage and algorithm design. It augments predictability and computability with an overarching stability principle. Stability expands on statistical uncertainty considerations to assess how human judgment calls impact data results through data and model/algorithm perturbations. As part of the PCS workflow, we develop PCS inference procedures, namely PCS perturbation intervals and PCS hypothesis testing, to investigate the stability of data results relative to problem formulation, data cleaning, modeling decisions, and interpretations. We illustrate PCS inference through neuroscience and genomics projects of our own and others. Moreover, we demonstrate its favorable performance over existing methods in terms of receiver operating characteristic (ROC) curves in high-dimensional, sparse linear model simulations, including a wide range of misspecified models. Finally, we propose PCS documentation based on R Markdown or Jupyter Notebook, with publicly available, reproducible codes and narratives to back up human choices made throughout an analysis. The PCS workflow and documentation are demonstrated in a genomics case study available on Zenodo.},
 	language = {en},
 	number = {8},
 	urldate = {2021-09-04},
@@ -14,17 +13,15 @@
 	month = feb,
 	year = {2020},
 	pages = {3920--3929},
-	file = {Yu and Kumbier - 2020 - Veridical data science.pdf:/home/james/Zotero/storage/2IMU9L5Q/Yu and Kumbier - 2020 - Veridical data science.pdf:application/pdf},
 }
 
 @article{lang-batchtools-2017,
-	title = {batchtools: {Tools} for {R} to work on batch systems},
+	title = {{batchtools: {Tools} for {R} to work on batch systems}},
 	volume = {2},
 	issn = {2475-9066},
 	shorttitle = {batchtools},
 	url = {https://joss.theoj.org/papers/10.21105/joss.00135},
 	doi = {10.21105/joss.00135},
-	abstract = {Lang et al, (2017), batchtools: Tools for R to work on batch systems, Journal of Open Source Software, 2(10), 135, doi:10.21105/joss.00135},
 	language = {en},
 	number = {10},
 	urldate = {2023-04-20},
@@ -33,7 +30,6 @@
 	month = feb,
 	year = {2017},
 	pages = {135},
-	file = {Full Text PDF:/home/james/Zotero/storage/PPA2UTP7/Lang et al. - 2017 - batchtools Tools for R to work on batch systems.pdf:application/pdf},
 }
 
 @article{wickham-welcome-2019,
@@ -42,7 +38,6 @@
 	issn = {2475-9066},
 	url = {https://joss.theoj.org/papers/10.21105/joss.01686},
 	doi = {10.21105/joss.01686},
-	abstract = {Wickham et al., (2019). Welcome to the Tidyverse. Journal of Open Source Software, 4(43), 1686, https://doi.org/10.21105/joss.01686},
 	language = {en},
 	number = {43},
 	urldate = {2023-04-20},
@@ -51,7 +46,6 @@
[…]
