### 2024-02-08 fc5fcba
Add 'Related R Packages' section to JOSS draft (#180)
See discussion in https://github.com/Yu-Group/simChef/issues/171
