### 2024-02-19 388226c
Document the Experiment R6 class
