+++ R/experiment.R
@@ -3,21 +3,36 @@ NULL
 
 #' `R6` class representing a simulation experiment.
 #'
+#' @name Experiment
+#'
 #' @docType class
 #'
-#' @description A simulation experiment with any number of `DGPs`,
-#'   `Methods`, `Evaluators`, and `Visualizers`.
+#' @description A simulation experiment with any number of [DGP],
+#'   [Method], [Evaluator], and [Visualizer] objects.
+#'
+#'   Generally speaking, users won't directly interact with the `Experiment` R6
+#'   class, but instead indirectly through [create_experiment()] and the tidy
+#'   `Experiment` helpers listed in below in the **See also** section.
 #'
 #' @details When run, an `Experiment` seamlessly combines `DGPs` and
 #'   `Methods`, computing results in parallel. Those results can then be
 #'   evaluated using `Evaluators` and visualized using `Visualizers`.
 #'
-#' @template experiment-template
+#' @seealso The following tidy helpers take an `Experiment` object as their
+#'   first argument: [create_experiment()], [generate_data()],
+#'   [fit_experiment()], [evaluate_experiment()], [visualize_experiment()],
+#'   [run_experiment()], [clear_cache()], [get_cached_results()],
+#'   [get_save_dir()], [set_save_dir()], [save_experiment()],
+#'   [export_visualizers()], [`add_*()`](add_funs.html),
+#'   [`update_*()`](update_funs.html), [`remove_*()`](remove_funs.html),
+#'   [`get_*()`](get_funs.html), and [`*_vary_across()`](vary_across.html).
 #'
 #' @export
 Experiment <- R6::R6Class(
   classname = 'Experiment',
   private = list(
+
+    # private fields
     .save_dir = NULL,
     .dgp_list = list(),
     .method_list = list(),
@@ -30,6 +45,7 @@ Experiment <- R6::R6Class(
     .fit_params = tibble::tibble(),
     .future.globals = TRUE,
     .future.packages = NULL,
+
     # private methods
     .add_obj = function(field_name, obj, obj_name, ...) {
       # TODO: check if obj is already in list by another name
@@ -56,6 +72,7 @@ Experiment <- R6::R6Class(
         private[[list_name]][[obj_name]] <- obj
       }
     },
+
     .update_obj = function(field_name, obj, obj_name, ...) {
       obj_list <- private$.get_obj_list(field_name, ...)
       if (!obj_name %in% names(obj_list)) {
@@ -71,6 +88,7 @@ Experiment <- R6::R6Class(
       list_name <- paste0(".", field_name, "_list")
       private[[list_name]][[obj_name]] <- obj
     },
+
     .remove_obj = function(field_name, obj_name, ...) {
       obj_list <- private$.get_obj_list(field_name, ...)
       list_name <- paste0(".", field_name, "_list")
@@ -89,6 +107,7 @@ Experiment <- R6::R6Class(
         private[[list_name]][[obj_name]] <- NULL
       }
     },
+
     .throw_empty_list_error = function(field_name, action_name = "run") {
       abort(
         sprintf(
@@ -99,11 +118,13 @@ Experiment <- R6::R6Class(
         call = rlang::caller_env()
       )
     },
+
     .get_obj_list = function(field_name, getter_name=NULL) {
       list_name <- paste0(".", field_name, "_list")
       obj_list <- private[[list_name]]
       return(obj_list)
     },
+
     .check_obj = function(obj, expected_class) {
       if (!inherits(obj, expected_class)) {
         err_msg <- sprintf("%s must be an instance of simChef::%s",
@@ -111,6 +132,7 @@ Experiment <- R6::R6Class(
         abort(err_msg, call = rlang::caller_env())
[…]
