+++ R/experiment-helpers.R
@@ -25,7 +25,7 @@
 #' @param n_reps The number of replicates of the `Experiment` for this run.
 #' @param parallel_strategy A vector with some combination of "reps", "dgps", or
 #'   "methods". Determines how computation will be distributed across available
-#'   resources. Default is "reps".
+#'   resources. Currently only the default, "reps", is supported.
 #' @param save If `TRUE`, save outputs to disk.
 #' @param use_cached Logical. If `TRUE`, find and return previously saved
 #'   results. If cached results cannot be found, continue as if `use_cached` was
@@ -198,7 +198,8 @@ create_experiment <- function(name = "experiment",
 #'   only the `n_reps` replicates are used when evaluating and visualizing
 #'   the `Experiment`.
 #'
-#' @return A list of results from the simulation experiment.
+#' @return A named list of results from the simulation experiment with the
+#'   following entries:
 #' \describe{
 #' \item{fit_results}{A tibble containing results from the `fit`
 #'   method. In addition to results columns, has columns named '.rep', '.dgp_name',
@@ -234,6 +235,10 @@ run_experiment <- function(experiment, n_reps = 1,
 #'
 #' @name generate_data
 #'
+#' @description Generate sample data from all `DGP` objects that were added
+#'   to the `Experiment`, including their varied params. Primarily useful
+#'   for debugging. Note that results are not generated in parallel.
+#'
 #' @inheritParams shared_experiment_helpers_args
 #' @param n_reps The number of datasets to generate per `DGP`.
 #'
@@ -251,11 +256,11 @@ run_experiment <- function(experiment, n_reps = 1,
 #' @examples
 #' # create DGP to generate data from normal distribution with n samples
 #' normal_dgp <- create_dgp(
-#'   .dgp_fun = function(n) rnorm(n), .name = "Normal DGP", n = 100
+#'   .dgp_fun = function(n) rnorm(n), .name = "Normal DGP", n = 10
 #' )
 #' # create DGP to generate data from binomial distribution with n samples
 #' bernoulli_dgp <- create_dgp(
-#'   .dgp_fun = function(n) rbinom(n, 1, 0.5), .name = "Bernoulli DGP", n = 100
+#'   .dgp_fun = function(n) rbinom(n, 1, 0.5), .name = "Bernoulli DGP", n = 10
 #' )
 #'
 #' # initialize experiment with toy DGPs only
@@ -265,7 +270,7 @@ run_experiment <- function(experiment, n_reps = 1,
 #'
 #' # generate data from all DGPs (and vary across components if applicable)
 #' # in experiment
-#' data_out <- generate_data(experiment, n_reps = 1)
+#' generate_data(experiment, n_reps = 2)
 #' @export
 generate_data <- function(experiment, n_reps=1, ...) {
   return(experiment$generate_data(n_reps = n_reps, ...))
@@ -348,6 +353,7 @@ visualize_experiment <- function(experiment, fit_results, eval_results = NULL,
                               verbose = verbose, ...))
 }
 
+# TODO: add @details
 #' Helper functions for adding components to an `Experiment`.
 #'
 #' @description Helper functions for adding `DGPs`, `Methods`,
@@ -463,6 +469,7 @@ add_visualizer <- function(experiment, visualizer, name=NULL, ...) {
   experiment$add_visualizer(visualizer, name, ...)
 }
 
+# TODO: add @details
 #' Helper functions for updating components of an `Experiment`.
 #'
 #' @description Helper functions for updating `DGPs`, `Methods`,
@@ -470,7 +477,7 @@ add_visualizer <- function(experiment, visualizer, name=NULL, ...) {
 #'   `Experiment`.
 #'
 #' @inheritParams add_funs
-#' @param name A name to identify the object to be updated.
[…]
