### 2024-02-20 b4136fc
Remove use of multisession in docs examples
Use of `future::plan("multisession", workers=10)` in the Evaluator and
Visualizer docs examples is likely causing issues in the check-r-package
job.
