+++ R/visualizer.R
@@ -74,7 +74,6 @@
 #'   add_evaluator(reject_prob_eval) %>%
 #'   add_vary_across(.dgp = dgp, rho = seq(0.91, 0.99, 0.02))
 #'
-#' future::plan("multisession", workers = 10)
 #' fit_results <- fit_experiment(experiment, n_reps=10)
 #' eval_results <- evaluate_experiment(experiment, fit_results)
 #'
@@ -234,7 +233,6 @@ create_visualizer <- function(.viz_fun, .name = NULL, .doc_options = list(),
 #'   add_evaluator(reject_prob_eval) %>%
 #'   add_vary_across(.dgp = dgp, rho = seq(0.91, 0.99, 0.02))
 #'
-#' future::plan("multisession", workers = 10)
 #' fit_results <- fit_experiment(experiment, n_reps=10)
 #' eval_results <- evaluate_experiment(experiment, fit_results)
 #'
