+++ man/create_evaluator.Rd
@@ -87,7 +87,6 @@ experiment <- create_experiment() \%>\%
   add_method(lm_method) \%>\%
   add_vary_across(.dgp = dgp, rho = seq(0.91, 0.99, 0.02))
 
-future::plan("multisession", workers = 10)
 fit_results <- fit_experiment(experiment, n_reps=10)
 
 # create an example Evaluator function
