+++ man/Visualizer.Rd
@@ -145,8 +145,8 @@ reject_prob_plot_fun <- function(eval_results, vary_params = NULL, eval_name) {
 }
 
 reject_prob_plot <- Visualizer$new(.viz_fun = reject_prob_plot_fun,
-                                   .name = "Rejection Prob. (alpha = 0.1) Plot",
-                                   eval_name = "Rejection Prob. (alpha = 0.1)")
+                                   .name = "Rejection Prob (alpha = 0.05) Plot",
+                                   eval_name = "Rejection Prob (alpha = 0.05)")
 
 reject_prob_plot$visualize(
   fit_results = fit_results, eval_results = eval_results, vary_params = "rho"
