### 2024-02-20 9983ab3
Improve documentation (#181)
* Addresses #174
* Add self-contained, executable example usage in readme
* Update concepts section of readme
* Showcase real world example in readme
* Add contributing guidelines
* Finish adding examples to docs
* Bug fix in add_na_counts
* Reorganize origins of simchef in readme
* Add simchef logo
* Some testing maintainence
* Re-document R6 classes

---------
Co-authored-by: James Duncan <jpdunc23@users.noreply.github.com>
Co-authored-by: James Duncan <jpduncan@berkeley.edu>
