+++ man-roxygen/dgp-template.R
@@ -1,13 +1,0 @@
-#' @section Public fields:
-#'
-#' \describe{
-#' \item{\code{name}}{The name of the \code{DGP}.}
-#' \item{\code{dgp_fun}}{The data-generating process function.}
-#' \item{\code{dgp_params}}{(Named) list of parameters to input into the data-generating process function.}
-#' }
-#'
-#' @section Public methods:
-#'
-#' \itemize{
-#' \item{\code{DGP$new()}} See \link{create_dgp}.
-#' }
