+++ man-roxygen/method-template.R
@@ -10,4 +10,5 @@
 #'
 #' \itemize{
 #' \item{\code{Method$new()}} See \link{create_method}.
+#' \item{\code{Method$fit()}} See \link{fit_method}.
 #' }

[…]
