### 2024-02-21 2034e19
fix yardstick::metric_summarizer() deprecation warning
* closes #182
