+++ R/evaluator-lib-utils.R
@@ -322,8 +322,8 @@ event_col <- function(xtab, event_level) {
   }
 }
 
-#' Helper function for constructing \code{finalize_esitmator_internal} for
-#'   custom \code{yardstick} metrics.
+#' Helper function for constructing `finalize_estimator_internal` for
+#'   custom `yardstick` metrics that are restricted to binary estimators.
 #' @keywords internal
 finalize_estimator_internal_constructor <- function(metric_dispatcher, x,
                                                     estimator) {
@@ -345,20 +345,24 @@ finalize_estimator_internal_constructor <- function(metric_dispatcher, x,
 #' Helper function for constructing \code{metric_vec} methods for
 #'   custom \code{yardstick} metrics.
 #' @keywords internal
-metric_vec_constructor <- function(name, fun, truth, estimate, estimator, na_rm,
-                                   event_level, ...) {
-  estimator <- yardstick::finalize_estimator(truth, estimator, 
-                                             metric_class = name)
-  
-  yardstick::metric_vec_template(
-    metric_impl = fun,
-    truth = truth,
-    estimate = estimate,
-    na_rm = na_rm,
-    cls = "factor",
-    estimator = estimator,
-    ...
+class_metric_vec_constructor <- function(name, fun, truth, estimate, estimator,
+                                         na_rm, case_weights, event_level,
+                                         ...) {
+  estimator <- yardstick::finalize_estimator(
+    truth, estimator, metric_class = name
   )
+  yardstick::check_class_metric(truth, estimate, case_weights, estimator)
+
+  if (na_rm) {
+    result <- yardstick::yardstick_remove_missing(truth, estimate, case_weights)
+    truth <- result$truth
+    estimate <- result$estimate
+    case_weights <- result$case_weights
+  } else if (yardstick::yardstick_any_missing(truth, estimate, case_weights)) {
+    return(NA_real_)
+  }
+
+  fun(truth, estimate, event_level)
 }
 
 #-------------------------- Custom Yardstick Metrics ---------------------------
@@ -384,37 +388,41 @@ tp <- function(data, ...) {
 }
 tp <- yardstick::new_class_metric(tp, direction = "maximize")
 
+#' @keywords internal
+tp_impl <- function(truth, estimate, event_level) {
+  xtab <- table(estimate, truth)
+  col <- event_col(xtab, event_level)
+  return(xtab[col, col])
+}
+
 #' @rdname tp
 #' @export
-tp.data.frame <- function(data, truth, estimate, estimator = NULL, 
-                          na_rm = FALSE, event_level = "first", ...) {
-  yardstick::metric_summarizer(
-    metric_nm = "tp",
-    metric_fn = tp_vec,
+tp.data.frame <- function(data, truth, estimate, estimator = NULL,
+                          na_rm = FALSE, case_weights = NULL,
+                          event_level = "first", ...) {
+  yardstick::class_metric_summarizer(
+    name = "tp",
+    fn = tp_vec,
     data = data,
     truth = !! rlang::enquo(truth),
     estimate = !! rlang::enquo(estimate), 
     estimator = estimator,
     na_rm = na_rm,
-    event_level = event_level,
-    ...
+    case_weights = !! rlang::enquo(case_weights),
+    event_level = event_level
   )
 }
 
 #' @rdname tp
 #' @export
-tp_vec <- function(truth, estimate, estimator = NULL, na_rm = FALSE, 
-                   event_level = "first", ...) {
-  
-  tp_impl <- function(truth, estimate) {
[…]
