### 2024-02-21 46846bc
update docs and fix dot arguments in visualizer library
* closes #163
