+++ man/pos.Rd
@@ -58,14 +58,13 @@ values should be stripped before the computation proceeds.}
 
 \item{case_weights}{The optional column identifier for case weights.
 This should be an unquoted column name that evaluates to a numeric column
-in \code{data}. For \verb{_vec()} functions, a numeric vector.}
+in \code{data}. For \verb{_vec()} functions, a numeric vector,
+\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}
 
 \item{event_level}{A single string. Either \code{"first"} or \code{"second"} to specify
 which level of \code{truth} to consider as the "event". This argument is only
-applicable when \code{estimator = "binary"}. The default uses an
-internal helper that generally defaults to \code{"first"}, however, if the
-deprecated global option \code{yardstick.event_first} is set, that will be
-used instead with a warning.}
+applicable when \code{estimator = "binary"}. The default uses an internal helper
+that defaults to \code{"first"}.}
 }
 \value{
 A \code{tibble} with columns \code{.metric}, \code{.estimator}, and
