### 2024-02-21 0ac4f5e
Library updates (#184)
* bug fix in `eval_contstructor()` when no arguments are provided through ...
* add violin plot option to `plot_eval_constructor()`

* fix yardstick::metric_summarizer() deprecation warning
* closes #182

* update docs and fix dot arguments in visualizer library
* closes #163
