### 2024-02-21 b4414f2
add option to show partial evaluator results in rmd output (#185)
* add `doc_nrows` field in `Evaluator`
* add `.doc_nrows` argument in `Evaluator$new()`
* add `nrows` argument to `set_doc_options`
* update docs and tests
