### 2024-02-27 Yu-Group/simChef#174 @jpdunc23
Hi @rcannood I just generated the new proof: https://github.com/openjournals/joss-papers/blob/joss.06156/joss.06156/10.21105.joss.06156.pdf
