### 2024-02-27 Yu-Group/simChef#175 @rcannood
> One point of confusion is the hpc argument which I think should be renamed / removed. Setting hpc = TRUE does nothing more than create directories called scripts and logs. We can definitely make this clearer for users.

I see! :+1: Sounds good!

---

>> Would it be possible to create separate articles in the documentation to showcase how to set up distributed computation, set up reproducibility, how to use caching & checkpointing, and how to debug runs? Or would you have an alternative solution?

> To clarify, do you think it would be better to created separate, self-contained vignettes for each concept? Another option is to split the comprehensive guide into smaller chunks (e.g. Part 1, Part 2, etc) focused on specific parts of the API. Would that achieve the same goal?

Indeed, I found the [simChef: a comphrensive guide](https://yu-group.github.io/simChef/articles/simChef-full.html) quite long, which makes it hard to find information regarding a specific aspect, e.g. caching or debugging. However, please consider this a non blocking issue as the information is definitely there.

---

For all of the other items I see you've created issues. I will also consider them optional improvements to the software tool, so that they are no longer blocking publication of the manuscript from my part :)
