### 2024-02-27 #6156 @rcannood
The authors made a lot of changes to the package, documentation and manuscript, thereby resolving all of my previous remarks in https://github.com/Yu-Group/simChef/issues/171, https://github.com/Yu-Group/simChef/issues/174 and https://github.com/Yu-Group/simChef/issues/175.

There are some minor improvements which are unresolved but I consider them non-blocking to the publication of this manuscript.

This means I checked off all of the items on my review checklist and would hereby like to recommend this manuscript for publication at JOSS :laughing:
