### 2024-03-13 1e0767f
fix duplicate vary_params bug (#188)
* If the same parameter is varies across different DGPs, only show this parameter once in `vary_params`
* Same applies to parameters that are varied across multiple methods
* Closes #187
