### 2024-03-15 9e223a1
Add DOIs and citation of MERITS paper
