+++ vignettes/paper.bib
@@ -78,17 +78,20 @@
     year =      {2020},
     volume =    {16},
     number =    {4},
-    url =       {http://www.tqmp.org/RegularArticles/vol16-4/p248/p248.pdf },
     pages =     {248-280},
-    doi =       {10.20982/tqmp.16.4.p248}
+    url =       {http://www.tqmp.org/RegularArticles/vol16-4/p248/p248.pdf },
+    doi =       {10.20982/tqmp.16.4.p248},
 }
 
-@Manual{kenny-SimEngine-2023,
-  title = {{SimEngine: A Modular Framework for Statistical Simulations in R}},
-  author = {Avi Kenny and Charles Wolock},
-  year = {2023},
-  note = {R package version 1.3.0},
-  url = {https://CRAN.R-project.org/package=SimEngine}
+@misc{kenny-simengine-2024,
+      title={{SimEngine: A Modular Framework for Statistical Simulations in R}},
+      author={Avi Kenny and Charles J. Wolock},
+      year={2024},
+      eprint={2403.05698},
+      archivePrefix={arXiv},
+      primaryClass={stat.CO},
+      url={https://doi.org/10.48550/arXiv.2403.05698},
+      doi={10.48550/arXiv.2403.05698},
 }
 
 @Manual{brown-simpr-2023,
@@ -99,8 +102,6 @@
 }
 
 @article{gasparini-rsimsum-2018,
-  doi = {10.21105/joss.00739},
-	url = {https://doi.org/10.21105/joss.00739},
 	year = {2018},
   publisher = {The Open Journal},
 	volume = {3},
@@ -108,7 +109,9 @@
 	pages = {739},
 	author = {Gasparini, Alessandro},
 	title = {{rsimsum: Summarise results from Monte Carlo simulation studies}},
-	journal = {Journal of Open Source Software}
+	journal = {Journal of Open Source Software},
+	url = {https://joss.theoj.org/papers/10.21105/joss.00739},
+  doi = {10.21105/joss.00739},
 }
 
 @article{blair-declaredesign-2019,
@@ -119,6 +122,8 @@
     Volume = {113},
     Number = {3},
     Pages = {838--859},
+    url = {https://doi.org/10.1017/S0003055419000194},
+    doi = {10.1017/S0003055419000194},
 }
 
 @Manual{joshi-simhelpers-2024,
@@ -168,6 +173,7 @@
   volume = {90},
   number = {2},
   pages = {1--30},
+  url = {https://dogi.org/10.18637/jss.v090.i02},
   doi = {10.18637/jss.v090.i02},
 }
 
@@ -179,6 +185,8 @@
   volume = {8},
   number = {3},
   pages = {704--713},
+  url = {https://dergipark.org.tr/en/download/article-file/1323860},
+  doi = {10.21449/ijate.804203},
 }
 
 @Misc{parsons-simMetric-2022,
@@ -198,7 +206,9 @@
       year={2016},
       eprint={1607.00021},
       archivePrefix={arXiv},
-      primaryClass={stat.CO}
+      primaryClass={stat.CO},
+      url={https://doi.org/10.48550/arXiv.1607.00021},
+      doi={10.48550/arXiv.1607.00021},
 }
 
 @Article{couch-infer-2021,
@@ -209,17 +219,28 @@
   volume = {6},
   number = {65},
   pages = {3661},
+  url = {https://joss.theoj.org/papers/10.21105/joss.03661},
   doi = {10.21105/joss.03661},
 }
 
 @article{hofert-simsalapar-2016,
  title={{Parallel and Other Simulations in R Made Easy: An End-to-End Study}},
  volume={69},
- url={https://www.jstatsoft.org/index.php/jss/article/view/v069i04},
[…]
