### 2024-03-15 6ba199f
Remove simChef hex logo from JOSS manuscript
