### 2024-03-15 5c3db9d
Add country to author affiliations in JOSS manuscript
