+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: simChef
 Title: Intensive Computational Experiments Made Easy
 Version: 0.1.0
 Authors@R: c(
-    person("Tiffany", "Tang", , "tiffany.tang@berkeley.edu", role = c("aut", "cre"),
+    person("Tiffany", "Tang", , "ttang4@nd.edu", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0002-8079-6867")),
     person("James", "Duncan", , "jpduncan@berkeley.edu", role = c("aut"),
            comment = c(ORCID = "0000-0003-3297-681X")),
