+++ DESCRIPTION
@@ -1,16 +1,17 @@
 Package: simChef
 Title: Intensive Computational Experiments Made Easy
-Version: 0.0.3
+Version: 0.1.0
 Authors@R: c(
-    person("James", "Duncan", , "jpduncan@berkeley.edu", role = c("aut", "cre"),
-           comment = c(ORCID = "0000-0003-3297-681X")),
-    person("Tiffany", "Tang", , "tiffany.tang@berkeley.edu", role = "aut",
+    person("Tiffany", "Tang", , "ttang4@nd.edu", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0002-8079-6867")),
+    person("James", "Duncan", , "jpduncan@berkeley.edu", role = c("aut"),
+           comment = c(ORCID = "0000-0003-3297-681X")),
     person("Corrine F.", "Elliott", role = "ctb",
            comment = c(ORCID = "0000-0001-7935-9945")),
     person("Philippe", "Boileau", role = "ctb",
            comment = c(ORCID = "0000-0002-4850-2507")),
-    person("Yu", "Bin", role = "ths")
+    person("Yu", "Bin", role = "ths",
+           comment = c(ORCID = "0000-0002-8888-4060"))
   )
 Description: Helps cook up the code necessary to create a fully-realized,
     high-quality, reproducible, and transparently documented simulation
