### 2024-03-20 #6156 @jpdunc23
Hi @Kevin-Mattheus-Moerman,

I've completed the post-review checklist, as well as the items below:

* [x]  Either remove the simchef logo from the paper, or use it as a proper figure with number/caption, and refer to it in the text.
* [x]  Check if you can add any potentially missing DOIs as per the above.
* [x]  Add the the country to all affiliations, e.g. `United States of America` (written like this in full).

Release 0.1.0 includes the latest changes from the review: https://github.com/Yu-Group/simChef/releases/tag/v0.1.0

Here is the DOI for the Zenodo archive for v0.1.0: [10.5281/zenodo.10845638](https://doi.org/10.5281/zenodo.10845638)

I think that should be everything but please let me know if I've overlooked anything. Thank you!
