### 2024-03-20 fd68fd6
Add DOI for MERITS paper
