+++ vignettes/paper.bib
@@ -243,4 +243,5 @@
       archivePrefix={arXiv},
       primaryClass={stat.CO},
       url={https://arxiv.org/abs/2403.08971},
+      doi={10.48550/arXiv.2403.08971},
 }
