+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: simChef
 Title: Intensive Computational Experiments Made Easy
-Version: 0.1.0
+Version: 0.1.0.9000
 Authors@R: c(
     person("Tiffany", "Tang", , "ttang4@nd.edu", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0002-8079-6867")),
