+++ dev/pkgdown.yml
@@ -1,11 +1,13 @@
-pandoc: 2.19.2
+pandoc: 3.1.11
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
+  cheat-sheet: cheat-sheet.html
   parallel: parallel.html
+  simChef-full: simChef-full.html
   simChef: simChef.html
   simulation-directory-setup: simulation-directory-setup.html
-last_built: 2023-07-03T16:38Z
+last_built: 2024-03-20T21:36Z
 urls:
   reference: https://yu-group.github.io/simChef/reference
   article: https://yu-group.github.io/simChef/articles
